LoopDataManager tests all passing
