--- LoopTests/Fixtures/flat_and_stable/flat_and_stable_counteraction_effect.json
@@ -1,230 +1,0 @@
-[
-  {
-    "startDate": "2020-08-11T17:25:13",
-    "endDate": "2020-08-11T17:30:13",
-    "unit": "mg\/min·dL",
-    "value": -0.17427698848393616
-  },
-  {
-    "startDate": "2020-08-11T17:30:13",
-    "endDate": "2020-08-11T17:35:13",
-    "unit": "mg\/min·dL",
-    "value": -0.172884893111717
-  },
-  {
-    "startDate": "2020-08-11T17:35:13",
-    "endDate": "2020-08-11T17:40:13",
-    "unit": "mg\/min·dL",
-    "value": -0.16620062119698026
-  },
-  {
-    "startDate": "2020-08-11T17:40:13",
-    "endDate": "2020-08-11T17:45:13",
-    "unit": "mg\/min·dL",
-    "value": -0.15239126546960888
-  },
-  {
-    "startDate": "2020-08-11T17:45:13",
-    "endDate": "2020-08-11T17:50:13",
-    "unit": "mg\/min·dL",
-    "value": -0.13844620387243192
-  },
-  {
-    "startDate": "2020-08-11T17:50:13",
-    "endDate": "2020-08-11T17:55:13",
-    "unit": "mg\/min·dL",
-    "value": -0.12440053903505803
-  },
-  {
-    "startDate": "2020-08-11T17:55:13",
-    "endDate": "2020-08-11T18:00:13",
-    "unit": "mg\/min·dL",
-    "value": -0.1102033787233404
-  },
-  {
-    "startDate": "2020-08-11T18:00:13",
-    "endDate": "2020-08-11T18:05:13",
-    "unit": "mg\/min·dL",
-    "value": -0.09582040633985235
-  },
-  {
-    "startDate": "2020-08-11T18:05:13",
-    "endDate": "2020-08-11T18:10:13",
-    "unit": "mg\/min·dL",
-    "value": -0.08123290693932182
-  },
-  {
-    "startDate": "2020-08-11T18:10:13",
-    "endDate": "2020-08-11T18:15:13",
-    "unit": "mg\/min·dL",
-    "value": -0.06643676319414542
-  },
-  {
-    "startDate": "2020-08-11T18:15:13",
-    "endDate": "2020-08-11T18:20:13",
-    "unit": "mg\/min·dL",
-    "value": -0.051441423013083416
-  },
-  {
-    "startDate": "2020-08-11T18:20:13",
-    "endDate": "2020-08-11T18:25:13",
-    "unit": "mg\/min·dL",
-    "value": -0.0362688411105418
-  },
-  {
-    "startDate": "2020-08-11T18:25:13",
-    "endDate": "2020-08-11T18:30:13",
-    "unit": "mg\/min·dL",
-    "value": -0.020952397377567107
-  },
-  {
-    "startDate": "2020-08-11T18:30:13",
-    "endDate": "2020-08-11T18:35:13",
-    "unit": "mg\/min·dL",
-    "value": -0.005535795415598254
-  },
-  {
-    "startDate": "2020-08-11T18:35:13",
-    "endDate": "2020-08-11T18:40:13",
-    "unit": "mg\/min·dL",
-    "value": 0.009928054942067454
-  },
-  {
-    "startDate": "2020-08-11T18:40:13",
-    "endDate": "2020-08-11T18:45:13",
-    "unit": "mg\/min·dL",
-    "value": 0.02537816688081129
-  },
-  {
-    "startDate": "2020-08-11T18:45:13",
-    "endDate": "2020-08-11T18:50:13",
-    "unit": "mg\/min·dL",
-    "value": 0.040746613021907935
-  },
-  {
-    "startDate": "2020-08-11T18:50:13",
-    "endDate": "2020-08-11T18:55:13",
-    "unit": "mg\/min·dL",
-    "value": 0.05595966408835151
-  },
-  {
-    "startDate": "2020-08-11T18:55:13",
-    "endDate": "2020-08-11T19:00:13",
-    "unit": "mg\/min·dL",
-    "value": 0.07093892464198123
-  },
-  {
-    "startDate": "2020-08-11T19:00:13",
-    "endDate": "2020-08-11T19:05:13",
-    "unit": "mg\/min·dL",
-    "value": 0.08560246050964196
-  },
-  {
-    "startDate": "2020-08-11T19:05:13",
-    "endDate": "2020-08-11T19:10:13",
-    "unit": "mg\/min·dL",
-    "value": 0.09986591236653002
-  },
-  {
-    "startDate": "2020-08-11T19:10:13",
-    "endDate": "2020-08-11T19:15:13",
-    "unit": "mg\/min·dL",
-    "value": 0.11364358985065513
-  },
-  {
-    "startDate": "2020-08-11T19:15:13",
-    "endDate": "2020-08-11T19:20:13",
-    "unit": "mg\/min·dL",
-    "value": 0.12684954054338973
-  },
-  {
-    "startDate": "2020-08-11T19:20:13",
-    "endDate": "2020-08-11T19:25:13",
-    "unit": "mg\/min·dL",
-    "value": 0.13939858816698666
-  },
-  {
-    "startDate": "2020-08-11T19:25:13",
-    "endDate": "2020-08-11T19:30:13",
-    "unit": "mg\/min·dL",
-    "value": 0.15120733442007542
-  },
-  {
-    "startDate": "2020-08-11T19:30:13",
-    "endDate": "2020-08-11T19:35:13",
-    "unit": "mg\/min·dL",
-    "value": 0.16219511899486355
-  },
-  {
-    "startDate": "2020-08-11T19:35:13",
-    "endDate": "2020-08-11T19:40:13",
-    "unit": "mg\/min·dL",
-    "value": 0.17228493249382398
-  },
-  {
-    "startDate": "2020-08-11T19:40:13",
-    "endDate": "2020-08-11T19:45:13",
-    "unit": "mg\/min·dL",
-    "value": 0.1814042771871964
-  },
-  {
-    "startDate": "2020-08-11T19:45:13",
-    "endDate": "2020-08-11T19:50:13",
-    "unit": "mg\/min·dL",
-    "value": 0.18948597082306992
-  },
-  {
-    "startDate": "2020-08-11T19:50:13",
-    "endDate": "2020-08-11T19:55:13",
-    "unit": "mg\/min·dL",
-    "value": 0.196468889016708
-  },
-  {
-    "startDate": "2020-08-11T19:55:13",
-    "endDate": "2020-08-11T20:00:13",
-    "unit": "mg\/min·dL",
-    "value": 0.20229864210263385
-  },
-  {
-    "startDate": "2020-08-11T20:00:13",
-    "endDate": "2020-08-11T20:05:13",
-    "unit": "mg\/min·dL",
-    "value": 0.2069281827278072
-  },
-  {
-    "startDate": "2020-08-11T20:05:13",
-    "endDate": "2020-08-11T20:10:13",
-    "unit": "mg\/min·dL",
-    "value": 0.21031834089428644
-  },
-  {
-    "startDate": "2020-08-11T20:10:13",
-    "endDate": "2020-08-11T20:15:13",
-    "unit": "mg\/min·dL",
-    "value": 0.21243828362120673
-  },
-  {
-    "startDate": "2020-08-11T20:15:13",
-    "endDate": "2020-08-11T20:20:13",
-    "unit": "mg\/min·dL",
-    "value": 0.213265896884441
-  },
-  {
-    "startDate": "2020-08-11T20:20:13",
-    "endDate": "2020-08-11T20:25:13",
-    "unit": "mg\/min·dL",
-    "value": 0.212788088004482
-  },
-  {
-    "startDate": "2020-08-11T20:25:13",
-    "endDate": "2020-08-11T20:32:50",
-    "unit": "mg\/min·dL",
-    "value": 0.17396858033976298
-  },
-  {
-    "startDate": "2020-08-11T20:32:50",
-    "endDate": "2020-08-11T20:45:02",
-    "unit": "mg\/min·dL",
-    "value": 0.18555611348135584
-  }
-]
