--- LoopTests/Fixtures/flat_and_stable/flat_and_stable_insulin_effect.json
@@ -1,377 +1,0 @@
-[
-  {
-    "date": "2020-08-11T20:50:00",
-    "amount": -0.21997829342610006,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T20:55:00",
-    "amount": -0.4261395410590354,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T21:00:00",
-    "amount": -0.7096583179105603,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T21:05:00",
-    "amount": -1.0621881093826662,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T21:10:00",
-    "amount": -1.4740341427597377,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T21:15:00",
-    "amount": -1.9363888584472242,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T21:20:00",
-    "amount": -2.441263560467393,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T21:25:00",
-    "amount": -2.9814248393095815,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T21:30:00",
-    "amount": -3.5503354629325354,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T21:35:00",
-    "amount": -4.142099441439137,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T21:40:00",
-    "amount": -4.751410989493849,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T21:45:00",
-    "amount": -5.373507127973413,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T21:50:00",
-    "amount": -6.004123682698768,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T21:55:00",
-    "amount": -6.639454453454031,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:00:00",
-    "amount": -7.276113340916081,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:05:00",
-    "amount": -7.911099232651796,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:10:00",
-    "amount": -8.541763462042216,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:15:00",
-    "amount": -9.165779665913185,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:20:00",
-    "amount": -9.7811158778376,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:25:00",
-    "amount": -10.386008704568662,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:30:00",
-    "amount": -10.97893944290868,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:35:00",
-    "amount": -11.558612003552255,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:40:00",
-    "amount": -12.12393251710345,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:45:00",
-    "amount": -12.673990505588074,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:50:00",
-    "amount": -13.20804151039699,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:55:00",
-    "amount": -13.725491074735217,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:00:00",
-    "amount": -14.225879985343203,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:05:00",
-    "amount": -14.708870684528089,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:10:00",
-    "amount": -15.174234769419765,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:15:00",
-    "amount": -15.62184150087279,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:20:00",
-    "amount": -16.05164724959357,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:25:00",
-    "amount": -16.463685811903716,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:30:00",
-    "amount": -16.858059532075337,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:35:00",
-    "amount": -17.234931172410647,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:40:00",
-    "amount": -17.594516476204813,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:45:00",
-    "amount": -17.93707737244358,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:50:00",
-    "amount": -18.26291577456192,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:55:00",
-    "amount": -18.572367928841064,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:00:00",
-    "amount": -18.86579927106296,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:05:00",
-    "amount": -19.14359975288623,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:10:00",
-    "amount": -19.406179602068192,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:15:00",
-    "amount": -19.65396548314523,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:20:00",
-    "amount": -19.887397027509305,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:25:00",
-    "amount": -20.106923703991654,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:30:00",
-    "amount": -20.313002003095814,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:35:00",
-    "amount": -20.506092909919293,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:40:00",
-    "amount": -20.686659642575187,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:45:00",
-    "amount": -20.855165634580125,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:50:00",
-    "amount": -21.012072741219335,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:55:00",
-    "amount": -21.157839651341906,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:00:00",
-    "amount": -21.292920487384542,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:05:00",
-    "amount": -21.41776357767731,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:10:00",
-    "amount": -21.532810386255537,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:15:00",
-    "amount": -21.638494586493437,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:20:00",
-    "amount": -21.735241265892345,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:25:00",
-    "amount": -21.823466250304694,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:30:00",
-    "amount": -21.903575536757817,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:35:00",
-    "amount": -21.975964824864416,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:40:00",
-    "amount": -22.041019137572135,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:45:00",
-    "amount": -22.099112522717494,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:50:00",
-    "amount": -22.150607827512605,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:55:00",
-    "amount": -22.19585653870966,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:00:00",
-    "amount": -22.235198681761787,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:05:00",
-    "amount": -22.268962772831713,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:10:00",
-    "amount": -22.297465817994798,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:15:00",
-    "amount": -22.32101335444279,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:20:00",
-    "amount": -22.33989952892162,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:25:00",
-    "amount": -22.354407209032342,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:30:00",
-    "amount": -22.364808123391935,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:35:00",
-    "amount": -22.371363026991066,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:40:00",
-    "amount": -22.374909853783546,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:45:00",
-    "amount": -22.37661999205696,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:50:00",
-    "amount": -22.377128476655095,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:55:00",
-    "amount": -22.377194743725912,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:00:00",
-    "amount": -22.37719474401739,
-    "unit": "mg/dL"
-  }
-]
