--- LoopTests/Fixtures/flat_and_stable/flat_and_stable_momentum_effect.json
@@ -1,1 +1,0 @@
-[]
