--- LoopTests/Fixtures/flat_and_stable/flat_and_stable_predicted_glucose.json
@@ -1,382 +1,0 @@
-[
-  {
-    "date": "2020-08-11T20:45:02",
-    "unit": "mg/dL",
-    "amount": 123.42849966275706
-  },
-  {
-    "date": "2020-08-11T20:50:00",
-    "unit": "mg/dL",
-    "amount": 124.26018046469977
-  },
-  {
-    "date": "2020-08-11T20:55:00",
-    "unit": "mg/dL",
-    "amount": 124.81009267337839
-  },
-  {
-    "date": "2020-08-11T21:00:00",
-    "unit": "mg/dL",
-    "amount": 125.20704000720727
-  },
-  {
-    "date": "2020-08-11T21:05:00",
-    "unit": "mg/dL",
-    "amount": 125.4593689807844
-  },
-  {
-    "date": "2020-08-11T21:10:00",
-    "unit": "mg/dL",
-    "amount": 125.57677436682542
-  },
-  {
-    "date": "2020-08-11T21:15:00",
-    "unit": "mg/dL",
-    "amount": 125.56806372492487
-  },
-  {
-    "date": "2020-08-11T21:20:00",
-    "unit": "mg/dL",
-    "amount": 125.44122575106047
-  },
-  {
-    "date": "2020-08-11T21:25:00",
-    "unit": "mg/dL",
-    "amount": 125.2034938547429
-  },
-  {
-    "date": "2020-08-11T21:30:00",
-    "unit": "mg/dL",
-    "amount": 124.86140526801341
-  },
-  {
-    "date": "2020-08-11T21:35:00",
-    "unit": "mg/dL",
-    "amount": 124.42085598076912
-  },
-  {
-    "date": "2020-08-11T21:40:00",
-    "unit": "mg/dL",
-    "amount": 123.88715177834555
-  },
-  {
-    "date": "2020-08-11T21:45:00",
-    "unit": "mg/dL",
-    "amount": 123.26505563986599
-  },
-  {
-    "date": "2020-08-11T21:50:00",
-    "unit": "mg/dL",
-    "amount": 122.63443908514064
-  },
-  {
-    "date": "2020-08-11T21:55:00",
-    "unit": "mg/dL",
-    "amount": 121.99910831438538
-  },
-  {
-    "date": "2020-08-11T22:00:00",
-    "unit": "mg/dL",
-    "amount": 121.36244942692333
-  },
-  {
-    "date": "2020-08-11T22:05:00",
-    "unit": "mg/dL",
-    "amount": 120.72746353518762
-  },
-  {
-    "date": "2020-08-11T22:10:00",
-    "unit": "mg/dL",
-    "amount": 120.0967993057972
-  },
-  {
-    "date": "2020-08-11T22:15:00",
-    "unit": "mg/dL",
-    "amount": 119.47278310192624
-  },
-  {
-    "date": "2020-08-11T22:20:00",
-    "unit": "mg/dL",
-    "amount": 118.85744689000182
-  },
-  {
-    "date": "2020-08-11T22:25:00",
-    "unit": "mg/dL",
-    "amount": 118.25255406327076
-  },
-  {
-    "date": "2020-08-11T22:30:00",
-    "unit": "mg/dL",
-    "amount": 117.65962332493075
-  },
-  {
-    "date": "2020-08-11T22:35:00",
-    "unit": "mg/dL",
-    "amount": 117.07995076428718
-  },
-  {
-    "date": "2020-08-11T22:40:00",
-    "unit": "mg/dL",
-    "amount": 116.51463025073598
-  },
-  {
-    "date": "2020-08-11T22:45:00",
-    "unit": "mg/dL",
-    "amount": 115.96457226225135
-  },
-  {
-    "date": "2020-08-11T22:50:00",
-    "unit": "mg/dL",
-    "amount": 115.43052125744244
-  },
-  {
-    "date": "2020-08-11T22:55:00",
-    "unit": "mg/dL",
-    "amount": 114.91307169310421
-  },
-  {
-    "date": "2020-08-11T23:00:00",
-    "unit": "mg/dL",
-    "amount": 114.41268278249623
-  },
-  {
-    "date": "2020-08-11T23:05:00",
-    "unit": "mg/dL",
-    "amount": 113.92969208331135
-  },
-  {
-    "date": "2020-08-11T23:10:00",
-    "unit": "mg/dL",
-    "amount": 113.46432799841968
-  },
-  {
-    "date": "2020-08-11T23:15:00",
-    "unit": "mg/dL",
-    "amount": 113.01672126696666
-  },
-  {
-    "date": "2020-08-11T23:20:00",
-    "unit": "mg/dL",
-    "amount": 112.58691551824587
-  },
-  {
-    "date": "2020-08-11T23:25:00",
-    "unit": "mg/dL",
-    "amount": 112.17487695593573
-  },
-  {
-    "date": "2020-08-11T23:30:00",
-    "unit": "mg/dL",
-    "amount": 111.78050323576412
-  },
-  {
-    "date": "2020-08-11T23:35:00",
-    "unit": "mg/dL",
-    "amount": 111.4036315954288
-  },
-  {
-    "date": "2020-08-11T23:40:00",
-    "unit": "mg/dL",
-    "amount": 111.04404629163464
-  },
-  {
-    "date": "2020-08-11T23:45:00",
-    "unit": "mg/dL",
-    "amount": 110.70148539539588
-  },
-  {
-    "date": "2020-08-11T23:50:00",
-    "unit": "mg/dL",
-    "amount": 110.37564699327754
-  },
-  {
-    "date": "2020-08-11T23:55:00",
-    "unit": "mg/dL",
-    "amount": 110.0661948389984
-  },
-  {
-    "date": "2020-08-12T00:00:00",
-    "unit": "mg/dL",
-    "amount": 109.7727634967765
-  },
-  {
-    "date": "2020-08-12T00:05:00",
-    "unit": "mg/dL",
-    "amount": 109.49496301495324
-  },
-  {
-    "date": "2020-08-12T00:10:00",
-    "unit": "mg/dL",
-    "amount": 109.23238316577128
-  },
-  {
-    "date": "2020-08-12T00:15:00",
-    "unit": "mg/dL",
-    "amount": 108.98459728469425
-  },
-  {
-    "date": "2020-08-12T00:20:00",
-    "unit": "mg/dL",
-    "amount": 108.75116574033018
-  },
-  {
-    "date": "2020-08-12T00:25:00",
-    "unit": "mg/dL",
-    "amount": 108.53163906384783
-  },
-  {
-    "date": "2020-08-12T00:30:00",
-    "unit": "mg/dL",
-    "amount": 108.32556076474367
-  },
-  {
-    "date": "2020-08-12T00:35:00",
-    "unit": "mg/dL",
-    "amount": 108.1324698579202
-  },
-  {
-    "date": "2020-08-12T00:40:00",
-    "unit": "mg/dL",
-    "amount": 107.95190312526431
-  },
-  {
-    "date": "2020-08-12T00:45:00",
-    "unit": "mg/dL",
-    "amount": 107.78339713325937
-  },
-  {
-    "date": "2020-08-12T00:50:00",
-    "unit": "mg/dL",
-    "amount": 107.62649002662016
-  },
-  {
-    "date": "2020-08-12T00:55:00",
-    "unit": "mg/dL",
-    "amount": 107.48072311649759
-  },
-  {
-    "date": "2020-08-12T01:00:00",
-    "unit": "mg/dL",
-    "amount": 107.34564228045495
-  },
-  {
-    "date": "2020-08-12T01:05:00",
-    "unit": "mg/dL",
-    "amount": 107.22079919016218
-  },
-  {
-    "date": "2020-08-12T01:10:00",
-    "unit": "mg/dL",
-    "amount": 107.10575238158395
-  },
-  {
-    "date": "2020-08-12T01:15:00",
-    "unit": "mg/dL",
-    "amount": 107.00006818134605
-  },
-  {
-    "date": "2020-08-12T01:20:00",
-    "unit": "mg/dL",
-    "amount": 106.90332150194715
-  },
-  {
-    "date": "2020-08-12T01:25:00",
-    "unit": "mg/dL",
-    "amount": 106.8150965175348
-  },
-  {
-    "date": "2020-08-12T01:30:00",
-    "unit": "mg/dL",
-    "amount": 106.73498723108167
-  },
-  {
-    "date": "2020-08-12T01:35:00",
-    "unit": "mg/dL",
-    "amount": 106.66259794297508
-  },
-  {
-    "date": "2020-08-12T01:40:00",
-    "unit": "mg/dL",
-    "amount": 106.59754363026737
-  },
-  {
-    "date": "2020-08-12T01:45:00",
-    "unit": "mg/dL",
-    "amount": 106.53945024512201
-  },
-  {
-    "date": "2020-08-12T01:50:00",
-    "unit": "mg/dL",
-    "amount": 106.4879549403269
-  },
-  {
-    "date": "2020-08-12T01:55:00",
-    "unit": "mg/dL",
-    "amount": 106.44270622912984
-  },
-  {
-    "date": "2020-08-12T02:00:00",
-    "unit": "mg/dL",
-    "amount": 106.40336408607772
-  },
-  {
-    "date": "2020-08-12T02:05:00",
-    "unit": "mg/dL",
-    "amount": 106.36959999500779
-  },
-  {
-    "date": "2020-08-12T02:10:00",
-    "unit": "mg/dL",
-    "amount": 106.34109694984471
-  },
-  {
-    "date": "2020-08-12T02:15:00",
-    "unit": "mg/dL",
-    "amount": 106.31754941339672
-  },
-  {
-    "date": "2020-08-12T02:20:00",
-    "unit": "mg/dL",
-    "amount": 106.2986632389179
-  },
-  {
-    "date": "2020-08-12T02:25:00",
-    "unit": "mg/dL",
-    "amount": 106.28415555880717
-  },
-  {
-    "date": "2020-08-12T02:30:00",
-    "unit": "mg/dL",
-    "amount": 106.27375464444758
-  },
-  {
-    "date": "2020-08-12T02:35:00",
-    "unit": "mg/dL",
-    "amount": 106.26719974084844
-  },
-  {
-    "date": "2020-08-12T02:40:00",
-    "unit": "mg/dL",
-    "amount": 106.26365291405597
-  },
-  {
-    "date": "2020-08-12T02:45:00",
-    "unit": "mg/dL",
-    "amount": 106.26194277578256
-  },
-  {
-    "date": "2020-08-12T02:50:00",
-    "unit": "mg/dL",
-    "amount": 106.26143429118443
-  },
-  {
-    "date": "2020-08-12T02:55:00",
-    "unit": "mg/dL",
-    "amount": 106.26136802411361
-  },
-  {
-    "date": "2020-08-12T03:00:00",
-    "unit": "mg/dL",
-    "amount": 106.26136802382213
-  }
-]
