--- LoopTests/Fixtures/high_and_falling/high_and_falling_carb_effect.json
@@ -1,1 +1,0 @@
-[]
