--- LoopTests/Fixtures/high_and_falling/high_and_falling_counteraction_effect.json
@@ -1,236 +1,0 @@
-[
-  {
-    "startDate": "2020-08-11T19:44:58",
-    "endDate": "2020-08-11T19:49:58",
-    "unit": "mg\/min·dL",
-    "value": 0.0
-  },
-  {
-    "startDate": "2020-08-11T19:49:58",
-    "endDate": "2020-08-11T19:54:58",
-    "unit": "mg\/min·dL",
-    "value": 0.0
-  },
-  {
-    "startDate": "2020-08-11T19:54:58",
-    "endDate": "2020-08-11T19:59:58",
-    "unit": "mg\/min·dL",
-    "value": 0.06065363877984119
-  },
-  {
-    "startDate": "2020-08-11T19:59:58",
-    "endDate": "2020-08-11T20:04:58",
-    "unit": "mg\/min·dL",
-    "value": 0.1829111566180655
-  },
-  {
-    "startDate": "2020-08-11T20:04:58",
-    "endDate": "2020-08-11T20:09:58",
-    "unit": "mg\/min·dL",
-    "value": 0.29002744966453
-  },
-  {
-    "startDate": "2020-08-11T20:09:58",
-    "endDate": "2020-08-11T20:14:58",
-    "unit": "mg\/min·dL",
-    "value": 0.38321365736330676
-  },
-  {
-    "startDate": "2020-08-11T20:14:58",
-    "endDate": "2020-08-11T20:19:58",
-    "unit": "mg\/min·dL",
-    "value": 0.4637144729903035
-  },
-  {
-    "startDate": "2020-08-11T20:19:58",
-    "endDate": "2020-08-11T20:24:58",
-    "unit": "mg\/min·dL",
-    "value": 0.5326798223434369
-  },
-  {
-    "startDate": "2020-08-11T20:24:58",
-    "endDate": "2020-08-11T20:29:58",
-    "unit": "mg\/min·dL",
-    "value": 0.5911714460685378
-  },
-  {
-    "startDate": "2020-08-11T20:29:58",
-    "endDate": "2020-08-11T20:34:58",
-    "unit": "mg\/min·dL",
-    "value": 0.6401690515783915
-  },
-  {
-    "startDate": "2020-08-11T20:34:58",
-    "endDate": "2020-08-11T20:39:58",
-    "unit": "mg\/min·dL",
-    "value": 0.6805760615235243
-  },
-  {
-    "startDate": "2020-08-11T20:39:58",
-    "endDate": "2020-08-11T20:44:58",
-    "unit": "mg\/min·dL",
-    "value": 0.7132249841389473
-  },
-  {
-    "startDate": "2020-08-11T20:44:58",
-    "endDate": "2020-08-11T20:49:58",
-    "unit": "mg\/min·dL",
-    "value": 0.7388824292522805
-  },
-  {
-    "startDate": "2020-08-11T20:49:58",
-    "endDate": "2020-08-11T20:54:58",
-    "unit": "mg\/min·dL",
-    "value": 0.758253792292099
-  },
-  {
-    "startDate": "2020-08-11T20:54:58",
-    "endDate": "2020-08-11T20:59:58",
-    "unit": "mg\/min·dL",
-    "value": 0.7719876272734658
-  },
-  {
-    "startDate": "2020-08-11T20:59:58",
-    "endDate": "2020-08-11T21:04:58",
-    "unit": "mg\/min·dL",
-    "value": 0.7806797284574882
-  },
-  {
-    "startDate": "2020-08-11T21:04:58",
-    "endDate": "2020-08-11T21:09:58",
-    "unit": "mg\/min·dL",
-    "value": 0.7848769391771567
-  },
-  {
-    "startDate": "2020-08-11T21:09:58",
-    "endDate": "2020-08-11T21:14:58",
-    "unit": "mg\/min·dL",
-    "value": 0.7850807051888878
-  },
-  {
-    "startDate": "2020-08-11T21:14:58",
-    "endDate": "2020-08-11T21:19:58",
-    "unit": "mg\/min·dL",
-    "value": 0.7817503888440966
-  },
-  {
-    "startDate": "2020-08-11T21:19:58",
-    "endDate": "2020-08-11T21:24:58",
-    "unit": "mg\/min·dL",
-    "value": 0.7753063593735205
-  },
-  {
-    "startDate": "2020-08-11T21:24:58",
-    "endDate": "2020-08-11T21:29:58",
-    "unit": "mg\/min·dL",
-    "value": 0.7661328736349247
-  },
-  {
-    "startDate": "2020-08-11T21:29:58",
-    "endDate": "2020-08-11T21:34:58",
-    "unit": "mg\/min·dL",
-    "value": 0.7545807607898111
-  },
-  {
-    "startDate": "2020-08-11T21:34:58",
-    "endDate": "2020-08-11T21:39:58",
-    "unit": "mg\/min·dL",
-    "value": 0.7409699235419351
-  },
-  {
-    "startDate": "2020-08-11T21:39:58",
-    "endDate": "2020-08-11T21:44:58",
-    "unit": "mg\/min·dL",
-    "value": 0.7255916677884272
-  },
-  {
-    "startDate": "2020-08-11T21:44:58",
-    "endDate": "2020-08-11T21:49:58",
-    "unit": "mg\/min·dL",
-    "value": 0.7087108717986296
-  },
-  {
-    "startDate": "2020-08-11T21:49:58",
-    "endDate": "2020-08-11T21:54:58",
-    "unit": "mg\/min·dL",
-    "value": 0.6905680053447725
-  },
-  {
-    "startDate": "2020-08-11T21:54:58",
-    "endDate": "2020-08-11T21:59:58",
-    "unit": "mg\/min·dL",
-    "value": 0.6713810085591916
-  },
-  {
-    "startDate": "2020-08-11T21:59:58",
-    "endDate": "2020-08-11T22:04:58",
-    "unit": "mg\/min·dL",
-    "value": 0.6513470396824913
-  },
-  {
-    "startDate": "2020-08-11T22:04:58",
-    "endDate": "2020-08-11T22:09:58",
-    "unit": "mg\/min·dL",
-    "value": 0.6306441002936196
-  },
-  {
-    "startDate": "2020-08-11T22:09:58",
-    "endDate": "2020-08-11T22:14:58",
-    "unit": "mg\/min·dL",
-    "value": 0.6094325460745351
-  },
-  {
-    "startDate": "2020-08-11T22:14:58",
-    "endDate": "2020-08-11T22:19:58",
-    "unit": "mg\/min·dL",
-    "value": 0.5878564906558068
-  },
-  {
-    "startDate": "2020-08-11T22:19:58",
-    "endDate": "2020-08-11T22:24:58",
-    "unit": "mg\/min·dL",
-    "value": 0.566045109614535
-  },
-  {
-    "startDate": "2020-08-11T22:24:58",
-    "endDate": "2020-08-11T22:29:58",
-    "unit": "mg\/min·dL",
-    "value": 0.5441138512497218
-  },
-  {
-    "startDate": "2020-08-11T22:29:58",
-    "endDate": "2020-08-11T22:34:58",
-    "unit": "mg\/min·dL",
-    "value": 0.5221655603410653
-  },
-  {
-    "startDate": "2020-08-11T22:34:58",
-    "endDate": "2020-08-11T22:39:58",
-    "unit": "mg\/min·dL",
-    "value": 0.5002915207035925
-  },
-  {
-    "startDate": "2020-08-11T22:39:58",
-    "endDate": "2020-08-11T22:44:58",
-    "unit": "mg\/min·dL",
-    "value": 0.47857242198147665
-  },
-  {
-    "startDate": "2020-08-11T22:44:58",
-    "endDate": "2020-08-11T22:49:44",
-    "unit": "mg\/min·dL",
-    "value": 0.0
-  },
-  {
-    "startDate": "2020-08-11T22:49:44",
-    "endDate": "2020-08-11T22:54:44",
-    "unit": "mg\/min·dL",
-    "value": 0.0
-  },
-  {
-    "startDate": "2020-08-11T22:54:44",
-    "endDate": "2020-08-11T22:59:45",
-    "unit": "mg\/min·dL",
-    "value": 0.060537504513367056
-  }
-]
