--- LoopTests/Fixtures/high_and_falling/high_and_falling_insulin_effect.json
@@ -1,377 +1,0 @@
-[
-  {
-    "date": "2020-08-11T23:00:00",
-    "amount": -0.30324421735766016,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:05:00",
-    "amount": -1.2074805603814895,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:10:00",
-    "amount": -2.6198776769809875,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:15:00",
-    "amount": -4.465672057725821,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:20:00",
-    "amount": -6.685266802723275,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:25:00",
-    "amount": -9.224809473113943,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:30:00",
-    "amount": -12.03541189572141,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:35:00",
-    "amount": -15.072766324251951,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:40:00",
-    "amount": -18.296788509858903,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:45:00",
-    "amount": -21.671285910499947,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:50:00",
-    "amount": -25.16364937991473,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:55:00",
-    "amount": -28.744566781673353,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:00:00",
-    "amount": -32.38775707198973,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:05:00",
-    "amount": -36.069723487241404,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:10:00",
-    "amount": -39.7695245587422,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:15:00",
-    "amount": -43.46856175861266,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:20:00",
-    "amount": -47.150382656903005,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:25:00",
-    "amount": -50.8004985417413,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:30:00",
-    "amount": -54.40621552148487,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:35:00",
-    "amount": -57.956478190913245,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:40:00",
-    "amount": -61.44172500265972,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:45:00",
-    "amount": -64.85375454057544,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:50:00",
-    "amount": -68.1856019437701,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:55:00",
-    "amount": -71.43142477888769,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:00:00",
-    "amount": -74.58639770394838,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:05:00",
-    "amount": -77.64661531000066,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:10:00",
-    "amount": -80.60900256705762,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:15:00",
-    "amount": -83.47123233849673,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:20:00",
-    "amount": -86.23164946343942,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:25:00",
-    "amount": -88.88920093973691,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:30:00",
-    "amount": -91.44337177121069,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:35:00",
-    "amount": -93.89412607185396,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:40:00",
-    "amount": -96.24185304691466,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:45:00",
-    "amount": -98.4873174962681,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:50:00",
-    "amount": -100.63161450934751,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:55:00",
-    "amount": -102.67612804323775,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:00:00",
-    "amount": -104.62249309644574,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:05:00",
-    "amount": -106.47256121042342,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:10:00",
-    "amount": -108.22836904922634,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:15:00",
-    "amount": -109.89210982481272,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:20:00",
-    "amount": -111.46610735150391,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:25:00",
-    "amount": -112.95279252810269,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:30:00",
-    "amount": -114.35468206016674,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:35:00",
-    "amount": -115.67435924802191,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:40:00",
-    "amount": -116.91445667832986,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:45:00",
-    "amount": -118.07764066845148,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:50:00",
-    "amount": -119.16659732352176,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:55:00",
-    "amount": -120.18402007612107,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:00:00",
-    "amount": -121.13259858773439,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:05:00",
-    "amount": -122.0150088998796,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:10:00",
-    "amount": -122.83390473089393,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:15:00",
-    "amount": -123.59190982193347,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:20:00",
-    "amount": -124.29161124279706,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:25:00",
-    "amount": -124.93555357476642,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:30:00",
-    "amount": -125.52623389378984,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:35:00",
-    "amount": -126.06609748305398,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:40:00",
-    "amount": -126.55753420931575,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:45:00",
-    "amount": -127.00287550232932,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:50:00",
-    "amount": -127.4043918813229,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:55:00",
-    "amount": -127.76429097678248,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:00:00",
-    "amount": -128.08471599980103,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:05:00",
-    "amount": -128.36774461497714,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:10:00",
-    "amount": -128.61538817630728,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:15:00",
-    "amount": -128.8295912887364,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:20:00",
-    "amount": -129.0122316610227,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:25:00",
-    "amount": -129.16512021834833,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:30:00",
-    "amount": -129.29000144569122,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:35:00",
-    "amount": -129.38855393536335,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:40:00",
-    "amount": -129.46239111434534,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:45:00",
-    "amount": -129.51306212910382,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:50:00",
-    "amount": -129.54205286749004,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:55:00",
-    "amount": -129.5507870990832,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T05:00:00",
-    "amount": -129.54961066748092,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T05:05:00",
-    "amount": -129.54931273055175,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T05:10:00",
-    "amount": -129.54930222233963,
-    "unit": "mg/dL"
-  }
-]
