--- LoopTests/Fixtures/high_and_falling/high_and_falling_momentum_effect.json
@@ -1,27 +1,0 @@
-[
-  {
-    "date": "2020-08-11T22:55:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T23:00:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T23:05:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T23:10:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T23:15:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  }
-]
