--- LoopTests/Fixtures/high_and_falling/high_and_falling_predicted_glucose.json
@@ -1,382 +1,0 @@
-[
-  {
-    "date": "2020-08-11T22:59:45",
-    "unit": "mg/dL",
-    "amount": 200.0
-  },
-  {
-    "date": "2020-08-11T23:00:00",
-    "unit": "mg/dL",
-    "amount": 200.0
-  },
-  {
-    "date": "2020-08-11T23:05:00",
-    "unit": "mg/dL",
-    "amount": 200.0111032633726
-  },
-  {
-    "date": "2020-08-11T23:10:00",
-    "unit": "mg/dL",
-    "amount": 200.01924237216699
-  },
-  {
-    "date": "2020-08-11T23:15:00",
-    "unit": "mg/dL",
-    "amount": 199.63033966967689
-  },
-  {
-    "date": "2020-08-11T23:20:00",
-    "unit": "mg/dL",
-    "amount": 198.52739386494645
-  },
-  {
-    "date": "2020-08-11T23:25:00",
-    "unit": "mg/dL",
-    "amount": 196.9449788576418
-  },
-  {
-    "date": "2020-08-11T23:30:00",
-    "unit": "mg/dL",
-    "amount": 194.9319828209393
-  },
-  {
-    "date": "2020-08-11T23:35:00",
-    "unit": "mg/dL",
-    "amount": 192.53271350113278
-  },
-  {
-    "date": "2020-08-11T23:40:00",
-    "unit": "mg/dL",
-    "amount": 189.78725514706883
-  },
-  {
-    "date": "2020-08-11T23:45:00",
-    "unit": "mg/dL",
-    "amount": 186.73180030078979
-  },
-  {
-    "date": "2020-08-11T23:50:00",
-    "unit": "mg/dL",
-    "amount": 183.398958108556
-  },
-  {
-    "date": "2020-08-11T23:55:00",
-    "unit": "mg/dL",
-    "amount": 179.81804070679738
-  },
-  {
-    "date": "2020-08-12T00:00:00",
-    "unit": "mg/dL",
-    "amount": 176.174850416481
-  },
-  {
-    "date": "2020-08-12T00:05:00",
-    "unit": "mg/dL",
-    "amount": 172.49288400122933
-  },
-  {
-    "date": "2020-08-12T00:10:00",
-    "unit": "mg/dL",
-    "amount": 168.79308292972854
-  },
-  {
-    "date": "2020-08-12T00:15:00",
-    "unit": "mg/dL",
-    "amount": 165.09404572985807
-  },
-  {
-    "date": "2020-08-12T00:20:00",
-    "unit": "mg/dL",
-    "amount": 161.41222483156773
-  },
-  {
-    "date": "2020-08-12T00:25:00",
-    "unit": "mg/dL",
-    "amount": 157.76210894672943
-  },
-  {
-    "date": "2020-08-12T00:30:00",
-    "unit": "mg/dL",
-    "amount": 154.15639196698586
-  },
-  {
-    "date": "2020-08-12T00:35:00",
-    "unit": "mg/dL",
-    "amount": 150.6061292975575
-  },
-  {
-    "date": "2020-08-12T00:40:00",
-    "unit": "mg/dL",
-    "amount": 147.12088248581102
-  },
-  {
-    "date": "2020-08-12T00:45:00",
-    "unit": "mg/dL",
-    "amount": 143.7088529478953
-  },
-  {
-    "date": "2020-08-12T00:50:00",
-    "unit": "mg/dL",
-    "amount": 140.37700554470064
-  },
-  {
-    "date": "2020-08-12T00:55:00",
-    "unit": "mg/dL",
-    "amount": 137.13118270958304
-  },
-  {
-    "date": "2020-08-12T01:00:00",
-    "unit": "mg/dL",
-    "amount": 133.97620978452235
-  },
-  {
-    "date": "2020-08-12T01:05:00",
-    "unit": "mg/dL",
-    "amount": 130.91599217847008
-  },
-  {
-    "date": "2020-08-12T01:10:00",
-    "unit": "mg/dL",
-    "amount": 127.95360492141312
-  },
-  {
-    "date": "2020-08-12T01:15:00",
-    "unit": "mg/dL",
-    "amount": 125.091375149974
-  },
-  {
-    "date": "2020-08-12T01:20:00",
-    "unit": "mg/dL",
-    "amount": 122.33095802503131
-  },
-  {
-    "date": "2020-08-12T01:25:00",
-    "unit": "mg/dL",
-    "amount": 119.67340654873382
-  },
-  {
-    "date": "2020-08-12T01:30:00",
-    "unit": "mg/dL",
-    "amount": 117.11923571726004
-  },
-  {
-    "date": "2020-08-12T01:35:00",
-    "unit": "mg/dL",
-    "amount": 114.66848141661677
-  },
-  {
-    "date": "2020-08-12T01:40:00",
-    "unit": "mg/dL",
-    "amount": 112.32075444155608
-  },
-  {
-    "date": "2020-08-12T01:45:00",
-    "unit": "mg/dL",
-    "amount": 110.07528999220263
-  },
-  {
-    "date": "2020-08-12T01:50:00",
-    "unit": "mg/dL",
-    "amount": 107.93099297912322
-  },
-  {
-    "date": "2020-08-12T01:55:00",
-    "unit": "mg/dL",
-    "amount": 105.88647944523298
-  },
-  {
-    "date": "2020-08-12T02:00:00",
-    "unit": "mg/dL",
-    "amount": 103.940114392025
-  },
-  {
-    "date": "2020-08-12T02:05:00",
-    "unit": "mg/dL",
-    "amount": 102.09004627804731
-  },
-  {
-    "date": "2020-08-12T02:10:00",
-    "unit": "mg/dL",
-    "amount": 100.33423843924439
-  },
-  {
-    "date": "2020-08-12T02:15:00",
-    "unit": "mg/dL",
-    "amount": 98.67049766365801
-  },
-  {
-    "date": "2020-08-12T02:20:00",
-    "unit": "mg/dL",
-    "amount": 97.09650013696682
-  },
-  {
-    "date": "2020-08-12T02:25:00",
-    "unit": "mg/dL",
-    "amount": 95.60981496036804
-  },
-  {
-    "date": "2020-08-12T02:30:00",
-    "unit": "mg/dL",
-    "amount": 94.207925428304
-  },
-  {
-    "date": "2020-08-12T02:35:00",
-    "unit": "mg/dL",
-    "amount": 92.88824824044882
-  },
-  {
-    "date": "2020-08-12T02:40:00",
-    "unit": "mg/dL",
-    "amount": 91.64815081014088
-  },
-  {
-    "date": "2020-08-12T02:45:00",
-    "unit": "mg/dL",
-    "amount": 90.48496682001925
-  },
-  {
-    "date": "2020-08-12T02:50:00",
-    "unit": "mg/dL",
-    "amount": 89.39601016494898
-  },
-  {
-    "date": "2020-08-12T02:55:00",
-    "unit": "mg/dL",
-    "amount": 88.37858741234966
-  },
-  {
-    "date": "2020-08-12T03:00:00",
-    "unit": "mg/dL",
-    "amount": 87.43000890073634
-  },
-  {
-    "date": "2020-08-12T03:05:00",
-    "unit": "mg/dL",
-    "amount": 86.54759858859113
-  },
-  {
-    "date": "2020-08-12T03:10:00",
-    "unit": "mg/dL",
-    "amount": 85.7287027575768
-  },
-  {
-    "date": "2020-08-12T03:15:00",
-    "unit": "mg/dL",
-    "amount": 84.97069766653726
-  },
-  {
-    "date": "2020-08-12T03:20:00",
-    "unit": "mg/dL",
-    "amount": 84.27099624567367
-  },
-  {
-    "date": "2020-08-12T03:25:00",
-    "unit": "mg/dL",
-    "amount": 83.62705391370432
-  },
-  {
-    "date": "2020-08-12T03:30:00",
-    "unit": "mg/dL",
-    "amount": 83.0363735946809
-  },
-  {
-    "date": "2020-08-12T03:35:00",
-    "unit": "mg/dL",
-    "amount": 82.49651000541675
-  },
-  {
-    "date": "2020-08-12T03:40:00",
-    "unit": "mg/dL",
-    "amount": 82.00507327915498
-  },
-  {
-    "date": "2020-08-12T03:45:00",
-    "unit": "mg/dL",
-    "amount": 81.55973198614141
-  },
-  {
-    "date": "2020-08-12T03:50:00",
-    "unit": "mg/dL",
-    "amount": 81.15821560714784
-  },
-  {
-    "date": "2020-08-12T03:55:00",
-    "unit": "mg/dL",
-    "amount": 80.79831651168826
-  },
-  {
-    "date": "2020-08-12T04:00:00",
-    "unit": "mg/dL",
-    "amount": 80.4778914886697
-  },
-  {
-    "date": "2020-08-12T04:05:00",
-    "unit": "mg/dL",
-    "amount": 80.19486287349359
-  },
-  {
-    "date": "2020-08-12T04:10:00",
-    "unit": "mg/dL",
-    "amount": 79.94721931216345
-  },
-  {
-    "date": "2020-08-12T04:15:00",
-    "unit": "mg/dL",
-    "amount": 79.73301619973434
-  },
-  {
-    "date": "2020-08-12T04:20:00",
-    "unit": "mg/dL",
-    "amount": 79.55037582744804
-  },
-  {
-    "date": "2020-08-12T04:25:00",
-    "unit": "mg/dL",
-    "amount": 79.3974872701224
-  },
-  {
-    "date": "2020-08-12T04:30:00",
-    "unit": "mg/dL",
-    "amount": 79.27260604277951
-  },
-  {
-    "date": "2020-08-12T04:35:00",
-    "unit": "mg/dL",
-    "amount": 79.17405355310738
-  },
-  {
-    "date": "2020-08-12T04:40:00",
-    "unit": "mg/dL",
-    "amount": 79.1002163741254
-  },
-  {
-    "date": "2020-08-12T04:45:00",
-    "unit": "mg/dL",
-    "amount": 79.04954535936692
-  },
-  {
-    "date": "2020-08-12T04:50:00",
-    "unit": "mg/dL",
-    "amount": 79.02055462098069
-  },
-  {
-    "date": "2020-08-12T04:55:00",
-    "unit": "mg/dL",
-    "amount": 79.01182038938752
-  },
-  {
-    "date": "2020-08-12T05:00:00",
-    "unit": "mg/dL",
-    "amount": 79.01299682098981
-  },
-  {
-    "date": "2020-08-12T05:05:00",
-    "unit": "mg/dL",
-    "amount": 79.01329475791898
-  },
-  {
-    "date": "2020-08-12T05:10:00",
-    "unit": "mg/dL",
-    "amount": 79.0133052661311
-  }
-]
