--- LoopTests/Fixtures/high_and_rising_with_cob/high_and_rising_with_cob_carb_effect.json
@@ -1,322 +1,0 @@
-[
-  {
-    "date": "2020-08-11T21:15:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T21:20:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T21:25:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T21:30:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T21:35:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T21:40:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T21:45:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T21:50:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T21:55:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T22:00:00",
-    "unit": "mg/dL",
-    "amount": 0.5054689190453953
-  },
-  {
-    "date": "2020-08-11T22:05:00",
-    "unit": "mg/dL",
-    "amount": 2.033246696823173
-  },
-  {
-    "date": "2020-08-11T22:10:00",
-    "unit": "mg/dL",
-    "amount": 3.5610244746009507
-  },
-  {
-    "date": "2020-08-11T22:15:00",
-    "unit": "mg/dL",
-    "amount": 5.088802252378729
-  },
-  {
-    "date": "2020-08-11T22:20:00",
-    "unit": "mg/dL",
-    "amount": 6.616580030156507
-  },
-  {
-    "date": "2020-08-11T22:25:00",
-    "unit": "mg/dL",
-    "amount": 8.144357807934284
-  },
-  {
-    "date": "2020-08-11T22:30:00",
-    "unit": "mg/dL",
-    "amount": 9.672135585712061
-  },
-  {
-    "date": "2020-08-11T22:35:00",
-    "unit": "mg/dL",
-    "amount": 11.199913363489841
-  },
-  {
-    "date": "2020-08-11T22:40:00",
-    "unit": "mg/dL",
-    "amount": 12.727691141267618
-  },
-  {
-    "date": "2020-08-11T22:45:00",
-    "unit": "mg/dL",
-    "amount": 14.255468919045395
-  },
-  {
-    "date": "2020-08-11T22:50:00",
-    "unit": "mg/dL",
-    "amount": 15.783246696823173
-  },
-  {
-    "date": "2020-08-11T22:55:00",
-    "unit": "mg/dL",
-    "amount": 17.311024474600952
-  },
-  {
-    "date": "2020-08-11T23:00:00",
-    "unit": "mg/dL",
-    "amount": 18.83880225237873
-  },
-  {
-    "date": "2020-08-11T23:05:00",
-    "unit": "mg/dL",
-    "amount": 20.366580030156506
-  },
-  {
-    "date": "2020-08-11T23:10:00",
-    "unit": "mg/dL",
-    "amount": 21.89435780793428
-  },
-  {
-    "date": "2020-08-11T23:15:00",
-    "unit": "mg/dL",
-    "amount": 23.422135585712063
-  },
-  {
-    "date": "2020-08-11T23:20:00",
-    "unit": "mg/dL",
-    "amount": 24.949913363489838
-  },
-  {
-    "date": "2020-08-11T23:25:00",
-    "unit": "mg/dL",
-    "amount": 26.477691141267616
-  },
-  {
-    "date": "2020-08-11T23:30:00",
-    "unit": "mg/dL",
-    "amount": 28.00546891904539
-  },
-  {
-    "date": "2020-08-11T23:35:00",
-    "unit": "mg/dL",
-    "amount": 29.533246696823177
-  },
-  {
-    "date": "2020-08-11T23:40:00",
-    "unit": "mg/dL",
-    "amount": 31.061024474600952
-  },
-  {
-    "date": "2020-08-11T23:45:00",
-    "unit": "mg/dL",
-    "amount": 32.58880225237873
-  },
-  {
-    "date": "2020-08-11T23:50:00",
-    "unit": "mg/dL",
-    "amount": 34.116580030156506
-  },
-  {
-    "date": "2020-08-11T23:55:00",
-    "unit": "mg/dL",
-    "amount": 35.644357807934284
-  },
-  {
-    "date": "2020-08-12T00:00:00",
-    "unit": "mg/dL",
-    "amount": 37.17213558571207
-  },
-  {
-    "date": "2020-08-12T00:05:00",
-    "unit": "mg/dL",
-    "amount": 38.69991336348984
-  },
-  {
-    "date": "2020-08-12T00:10:00",
-    "unit": "mg/dL",
-    "amount": 40.22769114126762
-  },
-  {
-    "date": "2020-08-12T00:15:00",
-    "unit": "mg/dL",
-    "amount": 41.7554689190454
-  },
-  {
-    "date": "2020-08-12T00:20:00",
-    "unit": "mg/dL",
-    "amount": 43.28324669682318
-  },
-  {
-    "date": "2020-08-12T00:25:00",
-    "unit": "mg/dL",
-    "amount": 44.81102447460095
-  },
-  {
-    "date": "2020-08-12T00:30:00",
-    "unit": "mg/dL",
-    "amount": 46.33880225237873
-  },
-  {
-    "date": "2020-08-12T00:35:00",
-    "unit": "mg/dL",
-    "amount": 47.86658003015651
-  },
-  {
-    "date": "2020-08-12T00:40:00",
-    "unit": "mg/dL",
-    "amount": 49.394357807934284
-  },
-  {
-    "date": "2020-08-12T00:45:00",
-    "unit": "mg/dL",
-    "amount": 50.922135585712056
-  },
-  {
-    "date": "2020-08-12T00:50:00",
-    "unit": "mg/dL",
-    "amount": 52.44991336348984
-  },
-  {
-    "date": "2020-08-12T00:55:00",
-    "unit": "mg/dL",
-    "amount": 53.97769114126762
-  },
-  {
-    "date": "2020-08-12T01:00:00",
-    "unit": "mg/dL",
-    "amount": 55.50546891904539
-  },
-  {
-    "date": "2020-08-12T01:05:00",
-    "unit": "mg/dL",
-    "amount": 57.03324669682318
-  },
-  {
-    "date": "2020-08-12T01:10:00",
-    "unit": "mg/dL",
-    "amount": 58.56102447460095
-  },
-  {
-    "date": "2020-08-12T01:15:00",
-    "unit": "mg/dL",
-    "amount": 60.08880225237873
-  },
-  {
-    "date": "2020-08-12T01:20:00",
-    "unit": "mg/dL",
-    "amount": 61.6165800301565
-  },
-  {
-    "date": "2020-08-12T01:25:00",
-    "unit": "mg/dL",
-    "amount": 63.144357807934284
-  },
-  {
-    "date": "2020-08-12T01:30:00",
-    "unit": "mg/dL",
-    "amount": 64.67213558571206
-  },
-  {
-    "date": "2020-08-12T01:35:00",
-    "unit": "mg/dL",
-    "amount": 66.19991336348984
-  },
-  {
-    "date": "2020-08-12T01:40:00",
-    "unit": "mg/dL",
-    "amount": 67.72769114126763
-  },
-  {
-    "date": "2020-08-12T01:45:00",
-    "unit": "mg/dL",
-    "amount": 69.2554689190454
-  },
-  {
-    "date": "2020-08-12T01:50:00",
-    "unit": "mg/dL",
-    "amount": 70.78324669682317
-  },
-  {
-    "date": "2020-08-12T01:55:00",
-    "unit": "mg/dL",
-    "amount": 72.31102447460096
-  },
-  {
-    "date": "2020-08-12T02:00:00",
-    "unit": "mg/dL",
-    "amount": 73.83880225237873
-  },
-  {
-    "date": "2020-08-12T02:05:00",
-    "unit": "mg/dL",
-    "amount": 75.3665800301565
-  },
-  {
-    "date": "2020-08-12T02:10:00",
-    "unit": "mg/dL",
-    "amount": 76.89435780793428
-  },
-  {
-    "date": "2020-08-12T02:15:00",
-    "unit": "mg/dL",
-    "amount": 78.42213558571207
-  },
-  {
-    "date": "2020-08-12T02:20:00",
-    "unit": "mg/dL",
-    "amount": 79.94991336348984
-  },
-  {
-    "date": "2020-08-12T02:25:00",
-    "unit": "mg/dL",
-    "amount": 81.47769114126761
-  },
-  {
-    "date": "2020-08-12T02:30:00",
-    "unit": "mg/dL",
-    "amount": 82.5
-  }
-]
