--- LoopTests/Fixtures/high_and_rising_with_cob/high_and_rising_with_cob_insulin_effect.json
@@ -1,387 +1,0 @@
-[
-  {
-    "date": "2020-08-11T21:50:00",
-    "amount": -8.639981829288883,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T21:55:00",
-    "amount": -9.789850828431643,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:00:00",
-    "amount": -10.963763653811602,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:05:00",
-    "amount": -12.153219270860628,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:10:00",
-    "amount": -13.350959307658405,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:15:00",
-    "amount": -14.550659188660132,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:20:00",
-    "amount": -15.7467157330705,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:25:00",
-    "amount": -16.934186099027563,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:30:00",
-    "amount": -18.108731231758313,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:35:00",
-    "amount": -19.266563509430355,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:40:00",
-    "amount": -20.404398300145722,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:45:00",
-    "amount": -21.51940916202376,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:50:00",
-    "amount": -22.60918643567319,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:55:00",
-    "amount": -23.67169899462816,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:00:00",
-    "amount": -24.705258934584283,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:05:00",
-    "amount": -25.708488996579725,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:10:00",
-    "amount": -26.680292532680262,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:15:00",
-    "amount": -27.619825835301093,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:20:00",
-    "amount": -28.526472663081833,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:25:00",
-    "amount": -29.3998208072736,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:30:00",
-    "amount": -30.239640552942898,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:35:00",
-    "amount": -31.045864898988505,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:40:00",
-    "amount": -31.818571410045358,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:45:00",
-    "amount": -32.557965581850254,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:50:00",
-    "amount": -33.26436560960395,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:55:00",
-    "amount": -33.93818845631585,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:00:00",
-    "amount": -34.57993712509237,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:05:00",
-    "amount": -35.190189045857444,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:10:00",
-    "amount": -35.76958549310118,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:15:00",
-    "amount": -36.31882195696637,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:20:00",
-    "amount": -36.838639395326744,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:25:00",
-    "amount": -37.32981629950877,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:30:00",
-    "amount": -37.7931615109809,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:35:00",
-    "amount": -38.229507730702785,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:40:00",
-    "amount": -38.639705666908725,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:45:00",
-    "amount": -39.024618770914344,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:50:00",
-    "amount": -39.38511851409847,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:55:00",
-    "amount": -39.72208016254089,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:00:00",
-    "amount": -40.03637900890356,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:05:00",
-    "amount": -40.32888702404502,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:10:00",
-    "amount": -40.600469893564416,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:15:00",
-    "amount": -40.85198440699897,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:20:00",
-    "amount": -41.08427616975518,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:25:00",
-    "amount": -41.29817761005208,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:30:00",
-    "amount": -41.494506255204584,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:35:00",
-    "amount": -41.674063253484796,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:40:00",
-    "amount": -41.837632119579226,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:45:00",
-    "amount": -41.98597768331826,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:50:00",
-    "amount": -42.11984522289805,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:55:00",
-    "amount": -42.23995976525269,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:00:00",
-    "amount": -42.347025537572655,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:05:00",
-    "amount": -42.441725555209814,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:10:00",
-    "amount": -42.524721332367534,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:15:00",
-    "amount": -42.59665270305005,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:20:00",
-    "amount": -42.65813774074594,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:25:00",
-    "amount": -42.70977276624976,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:30:00",
-    "amount": -42.752132433888306,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:35:00",
-    "amount": -42.785769887219686,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:40:00",
-    "amount": -42.81180494139787,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:45:00",
-    "amount": -42.831680423307795,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:50:00",
-    "amount": -42.84629245946508,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:55:00",
-    "amount": -42.856651353619604,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:00:00",
-    "amount": -42.86358529644523,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:05:00",
-    "amount": -42.867860863240104,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:10:00",
-    "amount": -42.87013681511805,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:15:00",
-    "amount": -42.871030675309036,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:20:00",
-    "amount": -42.871120411104464,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:25:00",
-    "amount": -42.87094608563874,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:30:00",
-    "amount": -42.870799980845575,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:35:00",
-    "amount": -42.87068168789406,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:40:00",
-    "amount": -42.870589529145974,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:45:00",
-    "amount": -42.870521892591526,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:50:00",
-    "amount": -42.87047723091304,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:55:00",
-    "amount": -42.870454060415405,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:00:00",
-    "amount": -42.87044984787703,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:05:00",
-    "amount": -42.87044984787703,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:10:00",
-    "amount": -42.87044984787703,
-    "unit": "mg/dL"
-  }
-]
