--- LoopTests/Fixtures/high_and_rising_with_cob/high_and_rising_with_cob_momentum_effect.json
@@ -1,27 +1,0 @@
-[
-  {
-    "date": "2020-08-11T21:45:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T21:50:00",
-    "unit": "mg/dL",
-    "amount": 0.0596641
-  },
-  {
-    "date": "2020-08-11T21:55:00",
-    "unit": "mg/dL",
-    "amount": 0.233866
-  },
-  {
-    "date": "2020-08-11T22:00:00",
-    "unit": "mg/dL",
-    "amount": 0.408067
-  },
-  {
-    "date": "2020-08-11T22:05:00",
-    "unit": "mg/dL",
-    "amount": 0.582269
-  }
-]
