--- LoopTests/Fixtures/high_and_stable/high_and_stable_carb_effect.json
@@ -1,322 +1,0 @@
-[
-  {
-    "date": "2020-08-12T12:05:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-12T12:10:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-12T12:15:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-12T12:20:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-12T12:25:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-12T12:30:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-12T12:35:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-12T12:40:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-12T12:45:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-12T12:50:00",
-    "unit": "mg/dL",
-    "amount": 0.03198444727394316
-  },
-  {
-    "date": "2020-08-12T12:55:00",
-    "unit": "mg/dL",
-    "amount": 0.4486511139406098
-  },
-  {
-    "date": "2020-08-12T13:00:00",
-    "unit": "mg/dL",
-    "amount": 0.8653177806072766
-  },
-  {
-    "date": "2020-08-12T13:05:00",
-    "unit": "mg/dL",
-    "amount": 1.281984447273943
-  },
-  {
-    "date": "2020-08-12T13:10:00",
-    "unit": "mg/dL",
-    "amount": 1.6986511139406095
-  },
-  {
-    "date": "2020-08-12T13:15:00",
-    "unit": "mg/dL",
-    "amount": 2.1153177806072767
-  },
-  {
-    "date": "2020-08-12T13:20:00",
-    "unit": "mg/dL",
-    "amount": 2.5319844472739432
-  },
-  {
-    "date": "2020-08-12T13:25:00",
-    "unit": "mg/dL",
-    "amount": 2.9486511139406097
-  },
-  {
-    "date": "2020-08-12T13:30:00",
-    "unit": "mg/dL",
-    "amount": 3.3653177806072763
-  },
-  {
-    "date": "2020-08-12T13:35:00",
-    "unit": "mg/dL",
-    "amount": 3.7819844472739437
-  },
-  {
-    "date": "2020-08-12T13:40:00",
-    "unit": "mg/dL",
-    "amount": 4.19865111394061
-  },
-  {
-    "date": "2020-08-12T13:45:00",
-    "unit": "mg/dL",
-    "amount": 4.615317780607277
-  },
-  {
-    "date": "2020-08-12T13:50:00",
-    "unit": "mg/dL",
-    "amount": 5.031984447273943
-  },
-  {
-    "date": "2020-08-12T13:55:00",
-    "unit": "mg/dL",
-    "amount": 5.44865111394061
-  },
-  {
-    "date": "2020-08-12T14:00:00",
-    "unit": "mg/dL",
-    "amount": 5.865317780607277
-  },
-  {
-    "date": "2020-08-12T14:05:00",
-    "unit": "mg/dL",
-    "amount": 6.281984447273943
-  },
-  {
-    "date": "2020-08-12T14:10:00",
-    "unit": "mg/dL",
-    "amount": 6.69865111394061
-  },
-  {
-    "date": "2020-08-12T14:15:00",
-    "unit": "mg/dL",
-    "amount": 7.115317780607277
-  },
-  {
-    "date": "2020-08-12T14:20:00",
-    "unit": "mg/dL",
-    "amount": 7.531984447273944
-  },
-  {
-    "date": "2020-08-12T14:25:00",
-    "unit": "mg/dL",
-    "amount": 7.94865111394061
-  },
-  {
-    "date": "2020-08-12T14:30:00",
-    "unit": "mg/dL",
-    "amount": 8.365317780607278
-  },
-  {
-    "date": "2020-08-12T14:35:00",
-    "unit": "mg/dL",
-    "amount": 8.781984447273942
-  },
-  {
-    "date": "2020-08-12T14:40:00",
-    "unit": "mg/dL",
-    "amount": 9.19865111394061
-  },
-  {
-    "date": "2020-08-12T14:45:00",
-    "unit": "mg/dL",
-    "amount": 9.615317780607278
-  },
-  {
-    "date": "2020-08-12T14:50:00",
-    "unit": "mg/dL",
-    "amount": 10.031984447273944
-  },
-  {
-    "date": "2020-08-12T14:55:00",
-    "unit": "mg/dL",
-    "amount": 10.44865111394061
-  },
-  {
-    "date": "2020-08-12T15:00:00",
-    "unit": "mg/dL",
-    "amount": 10.865317780607276
-  },
-  {
-    "date": "2020-08-12T15:05:00",
-    "unit": "mg/dL",
-    "amount": 11.281984447273942
-  },
-  {
-    "date": "2020-08-12T15:10:00",
-    "unit": "mg/dL",
-    "amount": 11.69865111394061
-  },
-  {
-    "date": "2020-08-12T15:15:00",
-    "unit": "mg/dL",
-    "amount": 12.115317780607278
-  },
-  {
-    "date": "2020-08-12T15:20:00",
-    "unit": "mg/dL",
-    "amount": 12.531984447273942
-  },
-  {
-    "date": "2020-08-12T15:25:00",
-    "unit": "mg/dL",
-    "amount": 12.94865111394061
-  },
-  {
-    "date": "2020-08-12T15:30:00",
-    "unit": "mg/dL",
-    "amount": 13.365317780607276
-  },
-  {
-    "date": "2020-08-12T15:35:00",
-    "unit": "mg/dL",
-    "amount": 13.781984447273944
-  },
-  {
-    "date": "2020-08-12T15:40:00",
-    "unit": "mg/dL",
-    "amount": 14.19865111394061
-  },
-  {
-    "date": "2020-08-12T15:45:00",
-    "unit": "mg/dL",
-    "amount": 14.615317780607274
-  },
-  {
-    "date": "2020-08-12T15:50:00",
-    "unit": "mg/dL",
-    "amount": 15.031984447273942
-  },
-  {
-    "date": "2020-08-12T15:55:00",
-    "unit": "mg/dL",
-    "amount": 15.44865111394061
-  },
-  {
-    "date": "2020-08-12T16:00:00",
-    "unit": "mg/dL",
-    "amount": 15.865317780607276
-  },
-  {
-    "date": "2020-08-12T16:05:00",
-    "unit": "mg/dL",
-    "amount": 16.281984447273942
-  },
-  {
-    "date": "2020-08-12T16:10:00",
-    "unit": "mg/dL",
-    "amount": 16.698651113940613
-  },
-  {
-    "date": "2020-08-12T16:15:00",
-    "unit": "mg/dL",
-    "amount": 17.115317780607278
-  },
-  {
-    "date": "2020-08-12T16:20:00",
-    "unit": "mg/dL",
-    "amount": 17.531984447273942
-  },
-  {
-    "date": "2020-08-12T16:25:00",
-    "unit": "mg/dL",
-    "amount": 17.94865111394061
-  },
-  {
-    "date": "2020-08-12T16:30:00",
-    "unit": "mg/dL",
-    "amount": 18.365317780607278
-  },
-  {
-    "date": "2020-08-12T16:35:00",
-    "unit": "mg/dL",
-    "amount": 18.781984447273945
-  },
-  {
-    "date": "2020-08-12T16:40:00",
-    "unit": "mg/dL",
-    "amount": 19.19865111394061
-  },
-  {
-    "date": "2020-08-12T16:45:00",
-    "unit": "mg/dL",
-    "amount": 19.615317780607278
-  },
-  {
-    "date": "2020-08-12T16:50:00",
-    "unit": "mg/dL",
-    "amount": 20.031984447273942
-  },
-  {
-    "date": "2020-08-12T16:55:00",
-    "unit": "mg/dL",
-    "amount": 20.44865111394061
-  },
-  {
-    "date": "2020-08-12T17:00:00",
-    "unit": "mg/dL",
-    "amount": 20.865317780607278
-  },
-  {
-    "date": "2020-08-12T17:05:00",
-    "unit": "mg/dL",
-    "amount": 21.281984447273942
-  },
-  {
-    "date": "2020-08-12T17:10:00",
-    "unit": "mg/dL",
-    "amount": 21.69865111394061
-  },
-  {
-    "date": "2020-08-12T17:15:00",
-    "unit": "mg/dL",
-    "amount": 22.115317780607278
-  },
-  {
-    "date": "2020-08-12T17:20:00",
-    "unit": "mg/dL",
-    "amount": 22.5
-  }
-]
