--- LoopTests/Fixtures/high_and_stable/high_and_stable_insulin_effect.json
@@ -1,382 +1,0 @@
-[
-  {
-    "date": "2020-08-12T12:40:00",
-    "amount": 0.0,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T12:45:00",
-    "amount": 0.0,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T12:50:00",
-    "amount": -0.00010857088891486093,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T12:55:00",
-    "amount": -0.11764496465132551,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T13:00:00",
-    "amount": -0.43873902047529706,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T13:05:00",
-    "amount": -0.9379108424564665,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T13:10:00",
-    "amount": -1.5919285563573975,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T13:15:00",
-    "amount": -2.379638252059979,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T13:20:00",
-    "amount": -3.281805691343955,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T13:25:00",
-    "amount": -4.280969013729399,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T13:30:00",
-    "amount": -5.361301721085654,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T13:35:00",
-    "amount": -6.508485266770114,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T13:40:00",
-    "amount": -7.709590617387781,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T13:45:00",
-    "amount": -8.952968195018745,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T13:50:00",
-    "amount": -10.228145645097738,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T13:55:00",
-    "amount": -11.525732910191868,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T14:00:00",
-    "amount": -12.837334122842806,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T14:05:00",
-    "amount": -14.15546586154826,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T14:10:00",
-    "amount": -15.473481342970688,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T14:15:00",
-    "amount": -16.785500150695594,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T14:20:00",
-    "amount": -18.08634312642022,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T14:25:00",
-    "amount": -19.3714720734388,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T14:30:00",
-    "amount": -20.636933944795025,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T14:35:00",
-    "amount": -21.8793092095861,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T14:40:00",
-    "amount": -23.095664110708345,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T14:45:00",
-    "amount": -24.28350654591071,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T14:50:00",
-    "amount": -25.440745321443842,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T14:55:00",
-    "amount": -26.565652543928085,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T15:00:00",
-    "amount": -27.65682893137946,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T15:05:00",
-    "amount": -28.71317183868988,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T15:10:00",
-    "amount": -29.733845806315355,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T15:15:00",
-    "amount": -30.71825545353683,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T15:20:00",
-    "amount": -31.666020549476087,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T15:25:00",
-    "amount": -32.576953106120015,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T15:30:00",
-    "amount": -33.45103634797771,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T15:35:00",
-    "amount": -34.2884054227078,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T15:40:00",
-    "amount": -35.08932972614962,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T15:45:00",
-    "amount": -35.854196723707794,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T15:50:00",
-    "amount": -36.58349715801203,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T15:55:00",
-    "amount": -37.27781154023619,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T16:00:00",
-    "amount": -37.93779782944274,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T16:05:00",
-    "amount": -38.564180210852335,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T16:10:00",
-    "amount": -39.15773889005006,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T16:15:00",
-    "amount": -39.7193008258551,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T16:20:00",
-    "amount": -40.24973132992669,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T16:25:00",
-    "amount": -40.749926466175516,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T16:30:00",
-    "amount": -41.220806187721365,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T16:35:00",
-    "amount": -41.66330815350251,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T16:40:00",
-    "amount": -42.078382170721305,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T16:45:00",
-    "amount": -42.46698521311987,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T16:50:00",
-    "amount": -42.8300769686383,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T16:55:00",
-    "amount": -43.16861587332966,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T17:00:00",
-    "amount": -43.483555591507375,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T17:05:00",
-    "amount": -43.77584190499485,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T17:10:00",
-    "amount": -44.04640997704762,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T17:15:00",
-    "amount": -44.296181959036595,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T17:20:00",
-    "amount": -44.52606491033073,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T17:25:00",
-    "amount": -44.736949004006426,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T17:30:00",
-    "amount": -44.92970599305237,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T17:35:00",
-    "amount": -45.10518791363997,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T17:40:00",
-    "amount": -45.264226003800765,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T17:45:00",
-    "amount": -45.40762981750194,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T17:50:00",
-    "amount": -45.53618651564582,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T17:55:00",
-    "amount": -45.650660316948574,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T18:00:00",
-    "amount": -45.75179209298248,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T18:05:00",
-    "amount": -45.8402990929014,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T18:10:00",
-    "amount": -45.9168747845193,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T18:15:00",
-    "amount": -45.98218879947835,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T18:20:00",
-    "amount": -46.036886971235035,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T18:25:00",
-    "amount": -46.08159145551451,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T18:30:00",
-    "amount": -46.116900923736516,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T18:35:00",
-    "amount": -46.14339082071065,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T18:40:00",
-    "amount": -46.16161367863287,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T18:45:00",
-    "amount": -46.17209948009722,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T18:50:00",
-    "amount": -46.175359225273134,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T18:55:00",
-    "amount": -46.175359225273134,
-    "unit": "mg/dL"
-  }
-]
