--- LoopTests/Fixtures/high_and_stable/high_and_stable_momentum_effect.json
@@ -1,27 +1,0 @@
-[
-  {
-    "date": "2020-08-12T12:35:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-12T12:40:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-12T12:45:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-12T12:50:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-12T12:55:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  }
-]
