--- LoopTests/Fixtures/high_and_stable/high_and_stable_predicted_glucose.json
@@ -1,387 +1,0 @@
-[
-  {
-    "date": "2020-08-12T12:39:22",
-    "unit": "mg/dL",
-    "amount": 200.0
-  },
-  {
-    "date": "2020-08-12T12:40:00",
-    "unit": "mg/dL",
-    "amount": 200.0
-  },
-  {
-    "date": "2020-08-12T12:45:00",
-    "unit": "mg/dL",
-    "amount": 200.00001542044052
-  },
-  {
-    "date": "2020-08-12T12:50:00",
-    "unit": "mg/dL",
-    "amount": 200.0120908555042
-  },
-  {
-    "date": "2020-08-12T12:55:00",
-    "unit": "mg/dL",
-    "amount": 200.22415504165645
-  },
-  {
-    "date": "2020-08-12T13:00:00",
-    "unit": "mg/dL",
-    "amount": 200.31998733993237
-  },
-  {
-    "date": "2020-08-12T13:05:00",
-    "unit": "mg/dL",
-    "amount": 200.23770477384636
-  },
-  {
-    "date": "2020-08-12T13:10:00",
-    "unit": "mg/dL",
-    "amount": 200.00053921763583
-  },
-  {
-    "date": "2020-08-12T13:15:00",
-    "unit": "mg/dL",
-    "amount": 199.6296445814189
-  },
-  {
-    "date": "2020-08-12T13:20:00",
-    "unit": "mg/dL",
-    "amount": 199.14425510341582
-  },
-  {
-    "date": "2020-08-12T13:25:00",
-    "unit": "mg/dL",
-    "amount": 198.56183264410652
-  },
-  {
-    "date": "2020-08-12T13:30:00",
-    "unit": "mg/dL",
-    "amount": 197.8982037016217
-  },
-  {
-    "date": "2020-08-12T13:35:00",
-    "unit": "mg/dL",
-    "amount": 197.1676868226039
-  },
-  {
-    "date": "2020-08-12T13:40:00",
-    "unit": "mg/dL",
-    "amount": 196.3832481386529
-  },
-  {
-    "date": "2020-08-12T13:45:00",
-    "unit": "mg/dL",
-    "amount": 195.5565372276886
-  },
-  {
-    "date": "2020-08-12T13:50:00",
-    "unit": "mg/dL",
-    "amount": 194.69802644427628
-  },
-  {
-    "date": "2020-08-12T13:55:00",
-    "unit": "mg/dL",
-    "amount": 193.81710584584883
-  },
-  {
-    "date": "2020-08-12T14:00:00",
-    "unit": "mg/dL",
-    "amount": 192.92217129986454
-  },
-  {
-    "date": "2020-08-12T14:05:00",
-    "unit": "mg/dL",
-    "amount": 192.02070622782577
-  },
-  {
-    "date": "2020-08-12T14:10:00",
-    "unit": "mg/dL",
-    "amount": 191.11935741307002
-  },
-  {
-    "date": "2020-08-12T14:15:00",
-    "unit": "mg/dL",
-    "amount": 190.2240052720118
-  },
-  {
-    "date": "2020-08-12T14:20:00",
-    "unit": "mg/dL",
-    "amount": 189.33982896295385
-  },
-  {
-    "date": "2020-08-12T14:25:00",
-    "unit": "mg/dL",
-    "amount": 188.47136668260194
-  },
-  {
-    "date": "2020-08-12T14:30:00",
-    "unit": "mg/dL",
-    "amount": 187.62257147791237
-  },
-  {
-    "date": "2020-08-12T14:35:00",
-    "unit": "mg/dL",
-    "amount": 186.79686287978797
-  },
-  {
-    "date": "2020-08-12T14:40:00",
-    "unit": "mg/dL",
-    "amount": 185.9971746453324
-  },
-  {
-    "date": "2020-08-12T14:45:00",
-    "unit": "mg/dL",
-    "amount": 185.2259988767967
-  },
-  {
-    "date": "2020-08-12T14:50:00",
-    "unit": "mg/dL",
-    "amount": 184.48542676793022
-  },
-  {
-    "date": "2020-08-12T14:55:00",
-    "unit": "mg/dL",
-    "amount": 183.77718621211264
-  },
-  {
-    "date": "2020-08-12T15:00:00",
-    "unit": "mg/dL",
-    "amount": 183.10267649132794
-  },
-  {
-    "date": "2020-08-12T15:05:00",
-    "unit": "mg/dL",
-    "amount": 182.4630002506842
-  },
-  {
-    "date": "2020-08-12T15:10:00",
-    "unit": "mg/dL",
-    "amount": 181.85899294972538
-  },
-  {
-    "date": "2020-08-12T15:15:00",
-    "unit": "mg/dL",
-    "amount": 181.29124996917056
-  },
-  {
-    "date": "2020-08-12T15:20:00",
-    "unit": "mg/dL",
-    "amount": 180.76015153989798
-  },
-  {
-    "date": "2020-08-12T15:25:00",
-    "unit": "mg/dL",
-    "amount": 180.26588564992073
-  },
-  {
-    "date": "2020-08-12T15:30:00",
-    "unit": "mg/dL",
-    "amount": 179.80846907472971
-  },
-  {
-    "date": "2020-08-12T15:35:00",
-    "unit": "mg/dL",
-    "amount": 179.3877666666663
-  },
-  {
-    "date": "2020-08-12T15:40:00",
-    "unit": "mg/dL",
-    "amount": 179.00350902989112
-  },
-  {
-    "date": "2020-08-12T15:45:00",
-    "unit": "mg/dL",
-    "amount": 178.65530869899962
-  },
-  {
-    "date": "2020-08-12T15:50:00",
-    "unit": "mg/dL",
-    "amount": 178.34267493136204
-  },
-  {
-    "date": "2020-08-12T15:55:00",
-    "unit": "mg/dL",
-    "amount": 178.06502721580455
-  },
-  {
-    "date": "2020-08-12T16:00:00",
-    "unit": "mg/dL",
-    "amount": 177.82170759326468
-  },
-  {
-    "date": "2020-08-12T16:05:00",
-    "unit": "mg/dL",
-    "amount": 177.61199187852174
-  },
-  {
-    "date": "2020-08-12T16:10:00",
-    "unit": "mg/dL",
-    "amount": 177.43509986599068
-  },
-  {
-    "date": "2020-08-12T16:15:00",
-    "unit": "mg/dL",
-    "amount": 177.2902045968523
-  },
-  {
-    "date": "2020-08-12T16:20:00",
-    "unit": "mg/dL",
-    "amount": 177.1764407594474
-  },
-  {
-    "date": "2020-08-12T16:25:00",
-    "unit": "mg/dL",
-    "amount": 177.09291228986524
-  },
-  {
-    "date": "2020-08-12T16:30:00",
-    "unit": "mg/dL",
-    "amount": 177.03869923498607
-  },
-  {
-    "date": "2020-08-12T16:35:00",
-    "unit": "mg/dL",
-    "amount": 177.0128639358716
-  },
-  {
-    "date": "2020-08-12T16:40:00",
-    "unit": "mg/dL",
-    "amount": 177.01445658531946
-  },
-  {
-    "date": "2020-08-12T16:45:00",
-    "unit": "mg/dL",
-    "amount": 177.04252020958756
-  },
-  {
-    "date": "2020-08-12T16:50:00",
-    "unit": "mg/dL",
-    "amount": 177.0960951207358
-  },
-  {
-    "date": "2020-08-12T16:55:00",
-    "unit": "mg/dL",
-    "amount": 177.17422288271112
-  },
-  {
-    "date": "2020-08-12T17:00:00",
-    "unit": "mg/dL",
-    "amount": 177.27594983120008
-  },
-  {
-    "date": "2020-08-12T17:05:00",
-    "unit": "mg/dL",
-    "amount": 177.40033018437927
-  },
-  {
-    "date": "2020-08-12T17:10:00",
-    "unit": "mg/dL",
-    "amount": 177.54642877899317
-  },
-  {
-    "date": "2020-08-12T17:15:00",
-    "unit": "mg/dL",
-    "amount": 177.71332346367086
-  },
-  {
-    "date": "2020-08-12T17:20:00",
-    "unit": "mg/dL",
-    "amount": 177.86812273176946
-  },
-  {
-    "date": "2020-08-12T17:25:00",
-    "unit": "mg/dL",
-    "amount": 177.65723863809376
-  },
-  {
-    "date": "2020-08-12T17:30:00",
-    "unit": "mg/dL",
-    "amount": 177.4644816490478
-  },
-  {
-    "date": "2020-08-12T17:35:00",
-    "unit": "mg/dL",
-    "amount": 177.2889997284602
-  },
-  {
-    "date": "2020-08-12T17:40:00",
-    "unit": "mg/dL",
-    "amount": 177.1299616382994
-  },
-  {
-    "date": "2020-08-12T17:45:00",
-    "unit": "mg/dL",
-    "amount": 176.9865578245982
-  },
-  {
-    "date": "2020-08-12T17:50:00",
-    "unit": "mg/dL",
-    "amount": 176.85800112645433
-  },
-  {
-    "date": "2020-08-12T17:55:00",
-    "unit": "mg/dL",
-    "amount": 176.74352732515158
-  },
-  {
-    "date": "2020-08-12T18:00:00",
-    "unit": "mg/dL",
-    "amount": 176.64239554911768
-  },
-  {
-    "date": "2020-08-12T18:05:00",
-    "unit": "mg/dL",
-    "amount": 176.55388854919875
-  },
-  {
-    "date": "2020-08-12T18:10:00",
-    "unit": "mg/dL",
-    "amount": 176.47731285758084
-  },
-  {
-    "date": "2020-08-12T18:15:00",
-    "unit": "mg/dL",
-    "amount": 176.41199884262178
-  },
-  {
-    "date": "2020-08-12T18:20:00",
-    "unit": "mg/dL",
-    "amount": 176.3573006708651
-  },
-  {
-    "date": "2020-08-12T18:25:00",
-    "unit": "mg/dL",
-    "amount": 176.3125961865856
-  },
-  {
-    "date": "2020-08-12T18:30:00",
-    "unit": "mg/dL",
-    "amount": 176.2772867183636
-  },
-  {
-    "date": "2020-08-12T18:35:00",
-    "unit": "mg/dL",
-    "amount": 176.25079682138946
-  },
-  {
-    "date": "2020-08-12T18:40:00",
-    "unit": "mg/dL",
-    "amount": 176.23257396346725
-  },
-  {
-    "date": "2020-08-12T18:45:00",
-    "unit": "mg/dL",
-    "amount": 176.22208816200288
-  },
-  {
-    "date": "2020-08-12T18:50:00",
-    "unit": "mg/dL",
-    "amount": 176.21882841682697
-  },
-  {
-    "date": "2020-08-12T18:55:00",
-    "unit": "mg/dL",
-    "amount": 176.21882841682697
-  }
-]
