--- LoopTests/Fixtures/low_and_falling/low_and_falling_carb_effect.json
@@ -1,322 +1,0 @@
-[
-  {
-    "date": "2020-08-11T21:35:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T21:40:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T21:45:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T21:50:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T21:55:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T22:00:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T22:05:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T22:10:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T22:15:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T22:20:00",
-    "unit": "mg/dL",
-    "amount": 1.113814925485187
-  },
-  {
-    "date": "2020-08-11T22:25:00",
-    "unit": "mg/dL",
-    "amount": 2.641592703262965
-  },
-  {
-    "date": "2020-08-11T22:30:00",
-    "unit": "mg/dL",
-    "amount": 4.169370481040743
-  },
-  {
-    "date": "2020-08-11T22:35:00",
-    "unit": "mg/dL",
-    "amount": 5.697148258818521
-  },
-  {
-    "date": "2020-08-11T22:40:00",
-    "unit": "mg/dL",
-    "amount": 7.224926036596299
-  },
-  {
-    "date": "2020-08-11T22:45:00",
-    "unit": "mg/dL",
-    "amount": 8.752703814374076
-  },
-  {
-    "date": "2020-08-11T22:50:00",
-    "unit": "mg/dL",
-    "amount": 10.280481592151855
-  },
-  {
-    "date": "2020-08-11T22:55:00",
-    "unit": "mg/dL",
-    "amount": 11.808259369929631
-  },
-  {
-    "date": "2020-08-11T23:00:00",
-    "unit": "mg/dL",
-    "amount": 13.336037147707408
-  },
-  {
-    "date": "2020-08-11T23:05:00",
-    "unit": "mg/dL",
-    "amount": 14.863814925485187
-  },
-  {
-    "date": "2020-08-11T23:10:00",
-    "unit": "mg/dL",
-    "amount": 16.391592703262965
-  },
-  {
-    "date": "2020-08-11T23:15:00",
-    "unit": "mg/dL",
-    "amount": 17.919370481040744
-  },
-  {
-    "date": "2020-08-11T23:20:00",
-    "unit": "mg/dL",
-    "amount": 19.44714825881852
-  },
-  {
-    "date": "2020-08-11T23:25:00",
-    "unit": "mg/dL",
-    "amount": 20.974926036596298
-  },
-  {
-    "date": "2020-08-11T23:30:00",
-    "unit": "mg/dL",
-    "amount": 22.502703814374076
-  },
-  {
-    "date": "2020-08-11T23:35:00",
-    "unit": "mg/dL",
-    "amount": 24.030481592151855
-  },
-  {
-    "date": "2020-08-11T23:40:00",
-    "unit": "mg/dL",
-    "amount": 25.558259369929633
-  },
-  {
-    "date": "2020-08-11T23:45:00",
-    "unit": "mg/dL",
-    "amount": 27.086037147707408
-  },
-  {
-    "date": "2020-08-11T23:50:00",
-    "unit": "mg/dL",
-    "amount": 28.613814925485187
-  },
-  {
-    "date": "2020-08-11T23:55:00",
-    "unit": "mg/dL",
-    "amount": 30.141592703262965
-  },
-  {
-    "date": "2020-08-12T00:00:00",
-    "unit": "mg/dL",
-    "amount": 31.66937048104074
-  },
-  {
-    "date": "2020-08-12T00:05:00",
-    "unit": "mg/dL",
-    "amount": 33.197148258818515
-  },
-  {
-    "date": "2020-08-12T00:10:00",
-    "unit": "mg/dL",
-    "amount": 34.7249260365963
-  },
-  {
-    "date": "2020-08-12T00:15:00",
-    "unit": "mg/dL",
-    "amount": 36.25270381437407
-  },
-  {
-    "date": "2020-08-12T00:20:00",
-    "unit": "mg/dL",
-    "amount": 37.78048159215186
-  },
-  {
-    "date": "2020-08-12T00:25:00",
-    "unit": "mg/dL",
-    "amount": 39.30825936992963
-  },
-  {
-    "date": "2020-08-12T00:30:00",
-    "unit": "mg/dL",
-    "amount": 40.83603714770741
-  },
-  {
-    "date": "2020-08-12T00:35:00",
-    "unit": "mg/dL",
-    "amount": 42.36381492548519
-  },
-  {
-    "date": "2020-08-12T00:40:00",
-    "unit": "mg/dL",
-    "amount": 43.891592703262965
-  },
-  {
-    "date": "2020-08-12T00:45:00",
-    "unit": "mg/dL",
-    "amount": 45.419370481040744
-  },
-  {
-    "date": "2020-08-12T00:50:00",
-    "unit": "mg/dL",
-    "amount": 46.947148258818515
-  },
-  {
-    "date": "2020-08-12T00:55:00",
-    "unit": "mg/dL",
-    "amount": 48.47492603659629
-  },
-  {
-    "date": "2020-08-12T01:00:00",
-    "unit": "mg/dL",
-    "amount": 50.00270381437408
-  },
-  {
-    "date": "2020-08-12T01:05:00",
-    "unit": "mg/dL",
-    "amount": 51.53048159215186
-  },
-  {
-    "date": "2020-08-12T01:10:00",
-    "unit": "mg/dL",
-    "amount": 53.05825936992963
-  },
-  {
-    "date": "2020-08-12T01:15:00",
-    "unit": "mg/dL",
-    "amount": 54.58603714770741
-  },
-  {
-    "date": "2020-08-12T01:20:00",
-    "unit": "mg/dL",
-    "amount": 56.113814925485194
-  },
-  {
-    "date": "2020-08-12T01:25:00",
-    "unit": "mg/dL",
-    "amount": 57.641592703262965
-  },
-  {
-    "date": "2020-08-12T01:30:00",
-    "unit": "mg/dL",
-    "amount": 59.169370481040744
-  },
-  {
-    "date": "2020-08-12T01:35:00",
-    "unit": "mg/dL",
-    "amount": 60.697148258818515
-  },
-  {
-    "date": "2020-08-12T01:40:00",
-    "unit": "mg/dL",
-    "amount": 62.2249260365963
-  },
-  {
-    "date": "2020-08-12T01:45:00",
-    "unit": "mg/dL",
-    "amount": 63.75270381437407
-  },
-  {
-    "date": "2020-08-12T01:50:00",
-    "unit": "mg/dL",
-    "amount": 65.28048159215186
-  },
-  {
-    "date": "2020-08-12T01:55:00",
-    "unit": "mg/dL",
-    "amount": 66.80825936992963
-  },
-  {
-    "date": "2020-08-12T02:00:00",
-    "unit": "mg/dL",
-    "amount": 68.33603714770742
-  },
-  {
-    "date": "2020-08-12T02:05:00",
-    "unit": "mg/dL",
-    "amount": 69.86381492548519
-  },
-  {
-    "date": "2020-08-12T02:10:00",
-    "unit": "mg/dL",
-    "amount": 71.39159270326296
-  },
-  {
-    "date": "2020-08-12T02:15:00",
-    "unit": "mg/dL",
-    "amount": 72.91937048104073
-  },
-  {
-    "date": "2020-08-12T02:20:00",
-    "unit": "mg/dL",
-    "amount": 74.44714825881853
-  },
-  {
-    "date": "2020-08-12T02:25:00",
-    "unit": "mg/dL",
-    "amount": 75.9749260365963
-  },
-  {
-    "date": "2020-08-12T02:30:00",
-    "unit": "mg/dL",
-    "amount": 77.50270381437407
-  },
-  {
-    "date": "2020-08-12T02:35:00",
-    "unit": "mg/dL",
-    "amount": 79.03048159215186
-  },
-  {
-    "date": "2020-08-12T02:40:00",
-    "unit": "mg/dL",
-    "amount": 80.55825936992963
-  },
-  {
-    "date": "2020-08-12T02:45:00",
-    "unit": "mg/dL",
-    "amount": 82.08603714770742
-  },
-  {
-    "date": "2020-08-12T02:50:00",
-    "unit": "mg/dL",
-    "amount": 82.5
-  }
-]
