--- LoopTests/Fixtures/low_and_falling/low_and_falling_counteraction_effect.json
@@ -1,218 +1,0 @@
-[
-  {
-    "startDate": "2020-08-11T19:06:06",
-    "endDate": "2020-08-11T19:11:06",
-    "unit": "mg\/min·dL",
-    "value": -0.3485359639226971
-  },
-  {
-    "startDate": "2020-08-11T19:11:06",
-    "endDate": "2020-08-11T19:16:06",
-    "unit": "mg\/min·dL",
-    "value": -0.34571948711910916
-  },
-  {
-    "startDate": "2020-08-11T19:16:06",
-    "endDate": "2020-08-11T19:21:06",
-    "unit": "mg\/min·dL",
-    "value": -0.3110996208816001
-  },
-  {
-    "startDate": "2020-08-11T19:21:06",
-    "endDate": "2020-08-11T19:26:06",
-    "unit": "mg\/min·dL",
-    "value": -0.17115290442012446
-  },
-  {
-    "startDate": "2020-08-11T19:26:06",
-    "endDate": "2020-08-11T19:31:06",
-    "unit": "mg\/min·dL",
-    "value": -0.035078937546724906
-  },
-  {
-    "startDate": "2020-08-11T19:31:06",
-    "endDate": "2020-08-11T19:36:06",
-    "unit": "mg\/min·dL",
-    "value": 0.08735109214809061
-  },
-  {
-    "startDate": "2020-08-11T19:36:06",
-    "endDate": "2020-08-11T19:41:06",
-    "unit": "mg\/min·dL",
-    "value": 0.19746935782304254
-  },
-  {
-    "startDate": "2020-08-11T19:41:06",
-    "endDate": "2020-08-11T19:46:06",
-    "unit": "mg\/min·dL",
-    "value": 0.2964814415989495
-  },
-  {
-    "startDate": "2020-08-11T19:46:06",
-    "endDate": "2020-08-11T19:51:06",
-    "unit": "mg\/min·dL",
-    "value": 0.3854747645772338
-  },
-  {
-    "startDate": "2020-08-11T19:51:06",
-    "endDate": "2020-08-11T19:56:06",
-    "unit": "mg\/min·dL",
-    "value": 0.4654266535840445
-  },
-  {
-    "startDate": "2020-08-11T19:56:06",
-    "endDate": "2020-08-11T20:01:06",
-    "unit": "mg\/min·dL",
-    "value": 0.5372120677140927
-  },
-  {
-    "startDate": "2020-08-11T20:01:06",
-    "endDate": "2020-08-11T20:06:06",
-    "unit": "mg\/min·dL",
-    "value": 0.6016110049547307
-  },
-  {
-    "startDate": "2020-08-11T20:06:06",
-    "endDate": "2020-08-11T20:11:06",
-    "unit": "mg\/min·dL",
-    "value": 0.6593156065538323
-  },
-  {
-    "startDate": "2020-08-11T20:11:06",
-    "endDate": "2020-08-11T20:16:06",
-    "unit": "mg\/min·dL",
-    "value": 0.7109369743543738
-  },
-  {
-    "startDate": "2020-08-11T20:16:06",
-    "endDate": "2020-08-11T20:21:06",
-    "unit": "mg\/min·dL",
-    "value": 0.7570117140551543
-  },
-  {
-    "startDate": "2020-08-11T20:21:06",
-    "endDate": "2020-08-11T20:26:06",
-    "unit": "mg\/min·dL",
-    "value": 0.7980082152690883
-  },
-  {
-    "startDate": "2020-08-11T20:26:06",
-    "endDate": "2020-08-11T20:31:06",
-    "unit": "mg\/min·dL",
-    "value": 0.8343326773392674
-  },
-  {
-    "startDate": "2020-08-11T20:31:06",
-    "endDate": "2020-08-11T20:36:06",
-    "unit": "mg\/min·dL",
-    "value": 0.8663348881353556
-  },
-  {
-    "startDate": "2020-08-11T20:36:06",
-    "endDate": "2020-08-11T20:41:06",
-    "unit": "mg\/min·dL",
-    "value": 0.894313761489434
-  },
-  {
-    "startDate": "2020-08-11T20:41:06",
-    "endDate": "2020-08-11T20:46:06",
-    "unit": "mg\/min·dL",
-    "value": 0.9185226375377566
-  },
-  {
-    "startDate": "2020-08-11T20:46:06",
-    "endDate": "2020-08-11T20:51:06",
-    "unit": "mg\/min·dL",
-    "value": 0.9391743490118711
-  },
-  {
-    "startDate": "2020-08-11T20:51:06",
-    "endDate": "2020-08-11T20:56:06",
-    "unit": "mg\/min·dL",
-    "value": 0.9564460554651047
-  },
-  {
-    "startDate": "2020-08-11T20:56:06",
-    "endDate": "2020-08-11T21:01:06",
-    "unit": "mg\/min·dL",
-    "value": 0.9704838465264228
-  },
-  {
-    "startDate": "2020-08-11T21:01:06",
-    "endDate": "2020-08-11T21:06:06",
-    "unit": "mg\/min·dL",
-    "value": 0.9814071145378676
-  },
-  {
-    "startDate": "2020-08-11T21:06:06",
-    "endDate": "2020-08-11T21:11:06",
-    "unit": "mg\/min·dL",
-    "value": 0.9893126963505664
-  },
-  {
-    "startDate": "2020-08-11T21:11:06",
-    "endDate": "2020-08-11T21:16:06",
-    "unit": "mg\/min·dL",
-    "value": 0.9942787836220887
-  },
-  {
-    "startDate": "2020-08-11T21:16:06",
-    "endDate": "2020-08-11T21:21:06",
-    "unit": "mg\/min·dL",
-    "value": 0.9963686006690381
-  },
-  {
-    "startDate": "2020-08-11T21:21:06",
-    "endDate": "2020-08-11T21:26:06",
-    "unit": "mg\/min·dL",
-    "value": 0.9956338487771189
-  },
-  {
-    "startDate": "2020-08-11T21:26:06",
-    "endDate": "2020-08-11T21:31:06",
-    "unit": "mg\/min·dL",
-    "value": 0.9921179158496414
-  },
-  {
-    "startDate": "2020-08-11T21:31:06",
-    "endDate": "2020-08-11T21:36:06",
-    "unit": "mg\/min·dL",
-    "value": 0.9858588503769765
-  },
-  {
-    "startDate": "2020-08-11T21:36:06",
-    "endDate": "2020-08-11T21:41:06",
-    "unit": "mg\/min·dL",
-    "value": 0.9768920989266177
-  },
-  {
-    "startDate": "2020-08-11T21:41:06",
-    "endDate": "2020-08-11T21:46:06",
-    "unit": "mg\/min·dL",
-    "value": 0.965253006677156
-  },
-  {
-    "startDate": "2020-08-11T21:46:06",
-    "endDate": "2020-08-11T21:51:06",
-    "unit": "mg\/min·dL",
-    "value": 0.9509790809419911
-  },
-  {
-    "startDate": "2020-08-11T21:51:06",
-    "endDate": "2020-08-11T21:56:06",
-    "unit": "mg\/min·dL",
-    "value": 0.9341120181395608
-  },
-  {
-    "startDate": "2020-08-11T21:56:06",
-    "endDate": "2020-08-11T22:01:06",
-    "unit": "mg\/min·dL",
-    "value": 0.9146994952586709
-  },
-  {
-    "startDate": "2020-08-11T22:01:06",
-    "endDate": "2020-08-11T22:06:06",
-    "unit": "mg\/min·dL",
-    "value": 0.8927967275284316
-  }
-]
