--- LoopTests/Fixtures/low_and_falling/low_and_falling_insulin_effect.json
@@ -1,382 +1,0 @@
-[
-  {
-    "date": "2020-08-11T22:05:00",
-    "amount": 0.0,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:10:00",
-    "amount": 0.0,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:15:00",
-    "amount": 0.0,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:20:00",
-    "amount": -0.1458612769290415,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:25:00",
-    "amount": -0.9512697097060898,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:30:00",
-    "amount": -2.3842190211605305,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:35:00",
-    "amount": -4.364249056420911,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:40:00",
-    "amount": -6.818055744021179,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:45:00",
-    "amount": -9.678947529165939,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:50:00",
-    "amount": -12.88633950788323,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:55:00",
-    "amount": -16.385282799253694,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:00:00",
-    "amount": -20.126026847056842,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:05:00",
-    "amount": -24.06361248698291,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:10:00",
-    "amount": -28.157493751577,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:15:00",
-    "amount": -32.37118651282725,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:20:00",
-    "amount": -36.67194218227609,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:25:00",
-    "amount": -41.03044480117815,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:30:00",
-    "amount": -45.420529958992645,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:35:00",
-    "amount": -49.81892407778424,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:40:00",
-    "amount": -54.205002693305985,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:45:00",
-    "amount": -58.56056645101005,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:50:00",
-    "amount": -62.869633617321924,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:55:00",
-    "amount": -67.11824798354047,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:00:00",
-    "amount": -71.29430111199574,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:05:00",
-    "amount": -75.38736794189215,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:10:00",
-    "amount": -79.38855483585641,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:15:00",
-    "amount": -83.29035920784969,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:20:00",
-    "amount": -87.0865399290309,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:25:00",
-    "amount": -90.77199776059544,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:30:00",
-    "amount": -94.34266511177375,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:35:00",
-    "amount": -97.79540446725352,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:40:00",
-    "amount": -101.12791487147612,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:45:00",
-    "amount": -104.33864589772718,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:50:00",
-    "amount": -107.42671856785853,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:55:00",
-    "amount": -110.39185272399912,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:00:00",
-    "amount": -113.23430038688294,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:05:00",
-    "amount": -115.95478466657976,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:10:00",
-    "amount": -118.55444382058852,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:15:00",
-    "amount": -121.03478008156584,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:20:00",
-    "amount": -123.39761290252783,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:25:00",
-    "amount": -125.64503629128907,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:30:00",
-    "amount": -127.7793799282899,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:35:00",
-    "amount": -129.8031737829074,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:40:00",
-    "amount": -131.71911596293566,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:45:00",
-    "amount": -133.53004355024044,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:50:00",
-    "amount": -135.23890619272336,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:55:00",
-    "amount": -136.84874223874277,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:00:00",
-    "amount": -138.3626572151035,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:05:00",
-    "amount": -139.78380446371185,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:10:00",
-    "amount": -141.1153677650564,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:15:00",
-    "amount": -142.3605457888761,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:20:00",
-    "amount": -143.5225382237712,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:25:00",
-    "amount": -144.6045334481494,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:30:00",
-    "amount": -145.60969761482852,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:35:00",
-    "amount": -146.54116503087826,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:40:00",
-    "amount": -147.40202972292892,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:45:00",
-    "amount": -148.19533808623217,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:50:00",
-    "amount": -148.9240825232751,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:55:00",
-    "amount": -149.5911959847532,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:00:00",
-    "amount": -150.1995473322352,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:05:00",
-    "amount": -150.7519374479337,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:10:00",
-    "amount": -151.251096022658,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:15:00",
-    "amount": -151.69967895829902,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:20:00",
-    "amount": -152.1002663261011,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:25:00",
-    "amount": -152.45536082654326,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:30:00",
-    "amount": -152.76738670089628,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:35:00",
-    "amount": -153.03868904847147,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:40:00",
-    "amount": -153.2715335072446,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:45:00",
-    "amount": -153.4681062589482,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:50:00",
-    "amount": -153.63051432289012,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:55:00",
-    "amount": -153.76078610569454,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:00:00",
-    "amount": -153.86087217688896,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:05:00",
-    "amount": -153.9326462427885,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:10:00",
-    "amount": -153.97790629347384,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:15:00",
-    "amount": -153.99837589983005,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:20:00",
-    "amount": -154.0,
-    "unit": "mg/dL"
-  }
-]
