--- LoopTests/Fixtures/low_and_falling/low_and_falling_momentum_effect.json
@@ -1,27 +1,0 @@
-[
-  {
-    "date": "2020-08-11T22:05:00",
-    "unit": "mg/dL",
-    "amount": 0.0
-  },
-  {
-    "date": "2020-08-11T22:10:00",
-    "unit": "mg/dL",
-    "amount": 1.35325
-  },
-  {
-    "date": "2020-08-11T22:15:00",
-    "unit": "mg/dL",
-    "amount": 3.09052
-  },
-  {
-    "date": "2020-08-11T22:20:00",
-    "unit": "mg/dL",
-    "amount": 4.8278
-  },
-  {
-    "date": "2020-08-11T22:25:00",
-    "unit": "mg/dL",
-    "amount": 6.56507
-  }
-]
