--- LoopTests/Fixtures/low_and_falling/low_and_falling_predicted_glucose.json
@@ -1,382 +1,0 @@
-[
-  {
-    "date": "2020-08-11T22:06:06",
-    "unit": "mg/dL",
-    "amount": 75.10768374646841
-  },
-  {
-    "date": "2020-08-11T22:10:00",
-    "unit": "mg/dL",
-    "amount": 76.46093289895596
-  },
-  {
-    "date": "2020-08-11T22:15:00",
-    "unit": "mg/dL",
-    "amount": 79.04942397908675
-  },
-  {
-    "date": "2020-08-11T22:20:00",
-    "unit": "mg/dL",
-    "amount": 83.00725362848293
-  },
-  {
-    "date": "2020-08-11T22:25:00",
-    "unit": "mg/dL",
-    "amount": 87.52123075828584
-  },
-  {
-    "date": "2020-08-11T22:30:00",
-    "unit": "mg/dL",
-    "amount": 91.12697884165053
-  },
-  {
-    "date": "2020-08-11T22:35:00",
-    "unit": "mg/dL",
-    "amount": 93.68408625591766
-  },
-  {
-    "date": "2020-08-11T22:40:00",
-    "unit": "mg/dL",
-    "amount": 95.26585707255327
-  },
-  {
-    "date": "2020-08-11T22:45:00",
-    "unit": "mg/dL",
-    "amount": 95.93898284635277
-  },
-  {
-    "date": "2020-08-11T22:50:00",
-    "unit": "mg/dL",
-    "amount": 95.76404848128813
-  },
-  {
-    "date": "2020-08-11T22:55:00",
-    "unit": "mg/dL",
-    "amount": 94.7960028582787
-  },
-  {
-    "date": "2020-08-11T23:00:00",
-    "unit": "mg/dL",
-    "amount": 93.08459653354495
-  },
-  {
-    "date": "2020-08-11T23:05:00",
-    "unit": "mg/dL",
-    "amount": 90.67478867139667
-  },
-  {
-    "date": "2020-08-11T23:10:00",
-    "unit": "mg/dL",
-    "amount": 88.10868518458037
-  },
-  {
-    "date": "2020-08-11T23:15:00",
-    "unit": "mg/dL",
-    "amount": 85.4227702011079
-  },
-  {
-    "date": "2020-08-11T23:20:00",
-    "unit": "mg/dL",
-    "amount": 82.64979230943683
-  },
-  {
-    "date": "2020-08-11T23:25:00",
-    "unit": "mg/dL",
-    "amount": 79.81906746831255
-  },
-  {
-    "date": "2020-08-11T23:30:00",
-    "unit": "mg/dL",
-    "amount": 76.95676008827584
-  },
-  {
-    "date": "2020-08-11T23:35:00",
-    "unit": "mg/dL",
-    "amount": 74.08614374726203
-  },
-  {
-    "date": "2020-08-11T23:40:00",
-    "unit": "mg/dL",
-    "amount": 71.22784290951806
-  },
-  {
-    "date": "2020-08-11T23:45:00",
-    "unit": "mg/dL",
-    "amount": 68.40005692959177
-  },
-  {
-    "date": "2020-08-11T23:50:00",
-    "unit": "mg/dL",
-    "amount": 65.61876754105768
-  },
-  {
-    "date": "2020-08-11T23:55:00",
-    "unit": "mg/dL",
-    "amount": 62.8979309526169
-  },
-  {
-    "date": "2020-08-12T00:00:00",
-    "unit": "mg/dL",
-    "amount": 60.24965560193941
-  },
-  {
-    "date": "2020-08-12T00:05:00",
-    "unit": "mg/dL",
-    "amount": 57.684366549820766
-  },
-  {
-    "date": "2020-08-12T00:10:00",
-    "unit": "mg/dL",
-    "amount": 55.21095743363429
-  },
-  {
-    "date": "2020-08-12T00:15:00",
-    "unit": "mg/dL",
-    "amount": 52.836930839418784
-  },
-  {
-    "date": "2020-08-12T00:20:00",
-    "unit": "mg/dL",
-    "amount": 50.568527896015354
-  },
-  {
-    "date": "2020-08-12T00:25:00",
-    "unit": "mg/dL",
-    "amount": 48.41084784222859
-  },
-  {
-    "date": "2020-08-12T00:30:00",
-    "unit": "mg/dL",
-    "amount": 46.36795826882805
-  },
-  {
-    "date": "2020-08-12T00:35:00",
-    "unit": "mg/dL",
-    "amount": 44.442996691126055
-  },
-  {
-    "date": "2020-08-12T00:40:00",
-    "unit": "mg/dL",
-    "amount": 42.63826406468122
-  },
-  {
-    "date": "2020-08-12T00:45:00",
-    "unit": "mg/dL",
-    "amount": 40.955310816207955
-  },
-  {
-    "date": "2020-08-12T00:50:00",
-    "unit": "mg/dL",
-    "amount": 39.39501592385437
-  },
-  {
-    "date": "2020-08-12T00:55:00",
-    "unit": "mg/dL",
-    "amount": 37.95765954549155
-  },
-  {
-    "date": "2020-08-12T01:00:00",
-    "unit": "mg/dL",
-    "amount": 36.642989660385524
-  },
-  {
-    "date": "2020-08-12T01:05:00",
-    "unit": "mg/dL",
-    "amount": 35.45028315846649
-  },
-  {
-    "date": "2020-08-12T01:10:00",
-    "unit": "mg/dL",
-    "amount": 34.3784017822355
-  },
-  {
-    "date": "2020-08-12T01:15:00",
-    "unit": "mg/dL",
-    "amount": 33.42584329903596
-  },
-  {
-    "date": "2020-08-12T01:20:00",
-    "unit": "mg/dL",
-    "amount": 32.59078825585176
-  },
-  {
-    "date": "2020-08-12T01:25:00",
-    "unit": "mg/dL",
-    "amount": 31.871142644868286
-  },
-  {
-    "date": "2020-08-12T01:30:00",
-    "unit": "mg/dL",
-    "amount": 31.264576785645232
-  },
-  {
-    "date": "2020-08-12T01:35:00",
-    "unit": "mg/dL",
-    "amount": 30.768560708805495
-  },
-  {
-    "date": "2020-08-12T01:40:00",
-    "unit": "mg/dL",
-    "amount": 30.380396306555042
-  },
-  {
-    "date": "2020-08-12T01:45:00",
-    "unit": "mg/dL",
-    "amount": 30.09724649702804
-  },
-  {
-    "date": "2020-08-12T01:50:00",
-    "unit": "mg/dL",
-    "amount": 29.91616163232291
-  },
-  {
-    "date": "2020-08-12T01:55:00",
-    "unit": "mg/dL",
-    "amount": 29.834103364081273
-  },
-  {
-    "date": "2020-08-12T02:00:00",
-    "unit": "mg/dL",
-    "amount": 29.84796616549835
-  },
-  {
-    "date": "2020-08-12T02:05:00",
-    "unit": "mg/dL",
-    "amount": 29.95459669466777
-  },
-  {
-    "date": "2020-08-12T02:10:00",
-    "unit": "mg/dL",
-    "amount": 30.150811171100997
-  },
-  {
-    "date": "2020-08-12T02:15:00",
-    "unit": "mg/dL",
-    "amount": 30.433410925059064
-  },
-  {
-    "date": "2020-08-12T02:20:00",
-    "unit": "mg/dL",
-    "amount": 30.799196267941767
-  },
-  {
-    "date": "2020-08-12T02:25:00",
-    "unit": "mg/dL",
-    "amount": 31.244978821341334
-  },
-  {
-    "date": "2020-08-12T02:30:00",
-    "unit": "mg/dL",
-    "amount": 31.767592432439983
-  },
-  {
-    "date": "2020-08-12T02:35:00",
-    "unit": "mg/dL",
-    "amount": 32.36390279416804
-  },
-  {
-    "date": "2020-08-12T02:40:00",
-    "unit": "mg/dL",
-    "amount": 33.03081587989516
-  },
-  {
-    "date": "2020-08-12T02:45:00",
-    "unit": "mg/dL",
-    "amount": 33.765285294369704
-  },
-  {
-    "date": "2020-08-12T02:50:00",
-    "unit": "mg/dL",
-    "amount": 33.45050370961934
-  },
-  {
-    "date": "2020-08-12T02:55:00",
-    "unit": "mg/dL",
-    "amount": 32.783390248141245
-  },
-  {
-    "date": "2020-08-12T03:00:00",
-    "unit": "mg/dL",
-    "amount": 32.175038900659246
-  },
-  {
-    "date": "2020-08-12T03:05:00",
-    "unit": "mg/dL",
-    "amount": 31.622648784960745
-  },
-  {
-    "date": "2020-08-12T03:10:00",
-    "unit": "mg/dL",
-    "amount": 31.12349021023644
-  },
-  {
-    "date": "2020-08-12T03:15:00",
-    "unit": "mg/dL",
-    "amount": 30.674907274595427
-  },
-  {
-    "date": "2020-08-12T03:20:00",
-    "unit": "mg/dL",
-    "amount": 30.27431990679335
-  },
-  {
-    "date": "2020-08-12T03:25:00",
-    "unit": "mg/dL",
-    "amount": 29.919225406351188
-  },
-  {
-    "date": "2020-08-12T03:30:00",
-    "unit": "mg/dL",
-    "amount": 29.607199531998162
-  },
-  {
-    "date": "2020-08-12T03:35:00",
-    "unit": "mg/dL",
-    "amount": 29.335897184422976
-  },
-  {
-    "date": "2020-08-12T03:40:00",
-    "unit": "mg/dL",
-    "amount": 29.10305272564983
-  },
-  {
-    "date": "2020-08-12T03:45:00",
-    "unit": "mg/dL",
-    "amount": 28.906479973946233
-  },
-  {
-    "date": "2020-08-12T03:50:00",
-    "unit": "mg/dL",
-    "amount": 28.744071910004322
-  },
-  {
-    "date": "2020-08-12T03:55:00",
-    "unit": "mg/dL",
-    "amount": 28.61380012719991
-  },
-  {
-    "date": "2020-08-12T04:00:00",
-    "unit": "mg/dL",
-    "amount": 28.513714056005483
-  },
-  {
-    "date": "2020-08-12T04:05:00",
-    "unit": "mg/dL",
-    "amount": 28.441939990105936
-  },
-  {
-    "date": "2020-08-12T04:10:00",
-    "unit": "mg/dL",
-    "amount": 28.396679939420608
-  },
-  {
-    "date": "2020-08-12T04:15:00",
-    "unit": "mg/dL",
-    "amount": 28.376210333064392
-  },
-  {
-    "date": "2020-08-12T04:20:00",
-    "unit": "mg/dL",
-    "amount": 28.374586232894444
-  }
-]
