--- LoopTests/Fixtures/low_with_low_treatment/low_with_low_treatment_insulin_effect.json
@@ -1,377 +1,0 @@
-[
-  {
-    "date": "2020-08-11T22:25:00",
-    "amount": -0.9512697097060898,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:30:00",
-    "amount": -2.3813732447934624,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:35:00",
-    "amount": -4.341390140188103,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:40:00",
-    "amount": -6.753751683906663,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:45:00",
-    "amount": -9.55387357996081,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:50:00",
-    "amount": -12.683720606187977,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T22:55:00",
-    "amount": -16.090664681076284,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:00:00",
-    "amount": -19.72706463270244,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:05:00",
-    "amount": -23.549875518271012,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:10:00",
-    "amount": -27.520285545942553,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:15:00",
-    "amount": -31.60337877339881,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:20:00",
-    "amount": -35.76782187287874,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:25:00",
-    "amount": -39.98557336066623,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:30:00",
-    "amount": -44.23161379064487,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:35:00",
-    "amount": -48.48369550694377,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:40:00",
-    "amount": -52.72211064025665,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:45:00",
-    "amount": -56.92947611647066,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:50:00",
-    "amount": -61.090534525122315,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-11T23:55:00",
-    "amount": -65.19196976921322,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:00:00",
-    "amount": -69.22223648736112,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:05:00",
-    "amount": -73.17140230440528,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:10:00",
-    "amount": -77.03100202768849,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:15:00",
-    "amount": -80.79390296354336,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:20:00",
-    "amount": -84.45418058224833,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:25:00",
-    "amount": -88.00700381010158,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:30:00",
-    "amount": -91.44852927449627,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:35:00",
-    "amount": -94.77580387215212,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:40:00",
-    "amount": -97.98667507215376,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:45:00",
-    "amount": -101.07970840432662,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:50:00",
-    "amount": -104.05411161991226,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T00:55:00",
-    "amount": -106.9096650456303,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:00:00",
-    "amount": -109.64665768417882,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:05:00",
-    "amount": -112.26582864415883,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:10:00",
-    "amount": -114.7683135104363,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:15:00",
-    "amount": -117.15559529219412,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:20:00",
-    "amount": -119.42945961048726,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:25:00",
-    "amount": -121.59195381009803,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:30:00",
-    "amount": -123.64534970199563,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:35:00",
-    "amount": -125.592109662823,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:40:00",
-    "amount": -127.43485583665301,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:45:00",
-    "amount": -129.17634220185357,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:50:00",
-    "amount": -130.81942928235597,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T01:55:00",
-    "amount": -132.36706129800115,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:00:00",
-    "amount": -133.8222455630132,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:05:00",
-    "amount": -135.18803395508212,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:10:00",
-    "amount": -136.46750629008383,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:15:00",
-    "amount": -137.66375544918807,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:20:00",
-    "amount": -138.779874116044,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:25:00",
-    "amount": -139.81894299195267,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:30:00",
-    "amount": -140.78402036646978,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:35:00",
-    "amount": -141.67813292977746,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:40:00",
-    "amount": -142.50426772146503,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:45:00",
-    "amount": -143.2653651180992,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:50:00",
-    "amount": -143.9643127691786,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T02:55:00",
-    "amount": -144.60394039779732,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:00:00",
-    "amount": -145.18701538860697,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:05:00",
-    "amount": -145.71623909150875,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:10:00",
-    "amount": -146.19424377494204,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:15:00",
-    "amount": -146.62359016770122,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:20:00",
-    "amount": -147.00676553292078,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:25:00",
-    "amount": -147.3461822222548,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:30:00",
-    "amount": -147.64417666235195,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:35:00",
-    "amount": -147.90300872951764,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:40:00",
-    "amount": -148.12486147197743,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:45:00",
-    "amount": -148.3118411424277,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:50:00",
-    "amount": -148.46597750659902,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T03:55:00",
-    "amount": -148.58922439637678,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:00:00",
-    "amount": -148.68346047864273,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:05:00",
-    "amount": -148.75049021342636,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:10:00",
-    "amount": -148.79204497720696,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:15:00",
-    "amount": -148.8097843292894,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:20:00",
-    "amount": -148.80959176148318,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:25:00",
-    "amount": -148.80862975663214,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:30:00",
-    "amount": -148.8083823028405,
-    "unit": "mg/dL"
-  },
-  {
-    "date": "2020-08-12T04:35:00",
-    "amount": -148.80836238795683,
-    "unit": "mg/dL"
-  }
-]
