--- LoopTests/Fixtures/low_with_low_treatment/low_with_low_treatment_momentum_effect.json
@@ -1,1 +1,0 @@
-[]
