--- LoopTests/Fixtures/low_with_low_treatment/low_with_low_treatment_predicted_glucose.json
@@ -1,382 +1,0 @@
-[
-  {
-    "date": "2020-08-11T22:23:55",
-    "unit": "mg/dL",
-    "amount": 81.22399763523448
-  },
-  {
-    "date": "2020-08-11T22:25:00",
-    "unit": "mg/dL",
-    "amount": 87.005525216014
-  },
-  {
-    "date": "2020-08-11T22:30:00",
-    "unit": "mg/dL",
-    "amount": 89.28803182494407
-  },
-  {
-    "date": "2020-08-11T22:35:00",
-    "unit": "mg/dL",
-    "amount": 90.82214183694292
-  },
-  {
-    "date": "2020-08-11T22:40:00",
-    "unit": "mg/dL",
-    "amount": 96.95805885168919
-  },
-  {
-    "date": "2020-08-11T22:45:00",
-    "unit": "mg/dL",
-    "amount": 103.32620850089171
-  },
-  {
-    "date": "2020-08-11T22:50:00",
-    "unit": "mg/dL",
-    "amount": 109.14614978329723
-  },
-  {
-    "date": "2020-08-11T22:55:00",
-    "unit": "mg/dL",
-    "amount": 114.47051078041765
-  },
-  {
-    "date": "2020-08-11T23:00:00",
-    "unit": "mg/dL",
-    "amount": 119.34693266417625
-  },
-  {
-    "date": "2020-08-11T23:05:00",
-    "unit": "mg/dL",
-    "amount": 123.81846037736848
-  },
-  {
-    "date": "2020-08-11T23:10:00",
-    "unit": "mg/dL",
-    "amount": 127.92390571183377
-  },
-  {
-    "date": "2020-08-11T23:15:00",
-    "unit": "mg/dL",
-    "amount": 131.69818460989035
-  },
-  {
-    "date": "2020-08-11T23:20:00",
-    "unit": "mg/dL",
-    "amount": 135.1726303992993
-  },
-  {
-    "date": "2020-08-11T23:25:00",
-    "unit": "mg/dL",
-    "amount": 133.3211329127054
-  },
-  {
-    "date": "2020-08-11T23:30:00",
-    "unit": "mg/dL",
-    "amount": 130.60287026050452
-  },
-  {
-    "date": "2020-08-11T23:35:00",
-    "unit": "mg/dL",
-    "amount": 127.87856632198339
-  },
-  {
-    "date": "2020-08-11T23:40:00",
-    "unit": "mg/dL",
-    "amount": 125.16792896644829
-  },
-  {
-    "date": "2020-08-11T23:45:00",
-    "unit": "mg/dL",
-    "amount": 122.48834126801206
-  },
-  {
-    "date": "2020-08-11T23:50:00",
-    "unit": "mg/dL",
-    "amount": 119.85506063713818
-  },
-  {
-    "date": "2020-08-11T23:55:00",
-    "unit": "mg/dL",
-    "amount": 117.28140317082506
-  },
-  {
-    "date": "2020-08-12T00:00:00",
-    "unit": "mg/dL",
-    "amount": 114.77891423045493
-  },
-  {
-    "date": "2020-08-12T00:05:00",
-    "unit": "mg/dL",
-    "amount": 112.35752619118857
-  },
-  {
-    "date": "2020-08-12T00:10:00",
-    "unit": "mg/dL",
-    "amount": 110.02570424568313
-  },
-  {
-    "date": "2020-08-12T00:15:00",
-    "unit": "mg/dL",
-    "amount": 107.79058108760603
-  },
-  {
-    "date": "2020-08-12T00:20:00",
-    "unit": "mg/dL",
-    "amount": 105.65808124667883
-  },
-  {
-    "date": "2020-08-12T00:25:00",
-    "unit": "mg/dL",
-    "amount": 103.63303579660337
-  },
-  {
-    "date": "2020-08-12T00:30:00",
-    "unit": "mg/dL",
-    "amount": 101.71928810998646
-  },
-  {
-    "date": "2020-08-12T00:35:00",
-    "unit": "mg/dL",
-    "amount": 99.91979129010838
-  },
-  {
-    "date": "2020-08-12T00:40:00",
-    "unit": "mg/dL",
-    "amount": 98.23669786788452
-  },
-  {
-    "date": "2020-08-12T00:45:00",
-    "unit": "mg/dL",
-    "amount": 96.67144231348942
-  },
-  {
-    "date": "2020-08-12T00:50:00",
-    "unit": "mg/dL",
-    "amount": 95.22481687568158
-  },
-  {
-    "date": "2020-08-12T00:55:00",
-    "unit": "mg/dL",
-    "amount": 93.89704122774131
-  },
-  {
-    "date": "2020-08-12T01:00:00",
-    "unit": "mg/dL",
-    "amount": 92.68782636697057
-  },
-  {
-    "date": "2020-08-12T01:05:00",
-    "unit": "mg/dL",
-    "amount": 91.59643318476833
-  },
-  {
-    "date": "2020-08-12T01:10:00",
-    "unit": "mg/dL",
-    "amount": 90.62172609626865
-  },
-  {
-    "date": "2020-08-12T01:15:00",
-    "unit": "mg/dL",
-    "amount": 89.76222209228861
-  },
-  {
-    "date": "2020-08-12T01:20:00",
-    "unit": "mg/dL",
-    "amount": 89.01613555177325
-  },
-  {
-    "date": "2020-08-12T01:25:00",
-    "unit": "mg/dL",
-    "amount": 88.38141912994024
-  },
-  {
-    "date": "2020-08-12T01:30:00",
-    "unit": "mg/dL",
-    "amount": 87.85580101582045
-  },
-  {
-    "date": "2020-08-12T01:35:00",
-    "unit": "mg/dL",
-    "amount": 87.43681883277084
-  },
-  {
-    "date": "2020-08-12T01:40:00",
-    "unit": "mg/dL",
-    "amount": 87.1218504367186
-  },
-  {
-    "date": "2020-08-12T01:45:00",
-    "unit": "mg/dL",
-    "amount": 86.90814184929582
-  },
-  {
-    "date": "2020-08-12T01:50:00",
-    "unit": "mg/dL",
-    "amount": 86.79283254657122
-  },
-  {
-    "date": "2020-08-12T01:55:00",
-    "unit": "mg/dL",
-    "amount": 86.77297830870381
-  },
-  {
-    "date": "2020-08-12T02:00:00",
-    "unit": "mg/dL",
-    "amount": 86.84557182146952
-  },
-  {
-    "date": "2020-08-12T02:05:00",
-    "unit": "mg/dL",
-    "amount": 87.00756120717838
-  },
-  {
-    "date": "2020-08-12T02:10:00",
-    "unit": "mg/dL",
-    "amount": 87.25586664995447
-  },
-  {
-    "date": "2020-08-12T02:15:00",
-    "unit": "mg/dL",
-    "amount": 87.58739526862803
-  },
-  {
-    "date": "2020-08-12T02:20:00",
-    "unit": "mg/dL",
-    "amount": 87.99905437954988
-  },
-  {
-    "date": "2020-08-12T02:25:00",
-    "unit": "mg/dL",
-    "amount": 88.48776328141898
-  },
-  {
-    "date": "2020-08-12T02:30:00",
-    "unit": "mg/dL",
-    "amount": 89.05046368467964
-  },
-  {
-    "date": "2020-08-12T02:35:00",
-    "unit": "mg/dL",
-    "amount": 89.68412889914973
-  },
-  {
-    "date": "2020-08-12T02:40:00",
-    "unit": "mg/dL",
-    "amount": 90.38577188523993
-  },
-  {
-    "date": "2020-08-12T02:45:00",
-    "unit": "mg/dL",
-    "amount": 90.82979584402324
-  },
-  {
-    "date": "2020-08-12T02:50:00",
-    "unit": "mg/dL",
-    "amount": 90.13084819294383
-  },
-  {
-    "date": "2020-08-12T02:55:00",
-    "unit": "mg/dL",
-    "amount": 89.49122056432512
-  },
-  {
-    "date": "2020-08-12T03:00:00",
-    "unit": "mg/dL",
-    "amount": 88.90814557351547
-  },
-  {
-    "date": "2020-08-12T03:05:00",
-    "unit": "mg/dL",
-    "amount": 88.37892187061368
-  },
-  {
-    "date": "2020-08-12T03:10:00",
-    "unit": "mg/dL",
-    "amount": 87.9009171871804
-  },
-  {
-    "date": "2020-08-12T03:15:00",
-    "unit": "mg/dL",
-    "amount": 87.47157079442121
-  },
-  {
-    "date": "2020-08-12T03:20:00",
-    "unit": "mg/dL",
-    "amount": 87.08839542920165
-  },
-  {
-    "date": "2020-08-12T03:25:00",
-    "unit": "mg/dL",
-    "amount": 86.74897873986762
-  },
-  {
-    "date": "2020-08-12T03:30:00",
-    "unit": "mg/dL",
-    "amount": 86.45098429977048
-  },
-  {
-    "date": "2020-08-12T03:35:00",
-    "unit": "mg/dL",
-    "amount": 86.1921522326048
-  },
-  {
-    "date": "2020-08-12T03:40:00",
-    "unit": "mg/dL",
-    "amount": 85.97029949014501
-  },
-  {
-    "date": "2020-08-12T03:45:00",
-    "unit": "mg/dL",
-    "amount": 85.78331981969473
-  },
-  {
-    "date": "2020-08-12T03:50:00",
-    "unit": "mg/dL",
-    "amount": 85.62918345552342
-  },
-  {
-    "date": "2020-08-12T03:55:00",
-    "unit": "mg/dL",
-    "amount": 85.50593656574566
-  },
-  {
-    "date": "2020-08-12T04:00:00",
-    "unit": "mg/dL",
-    "amount": 85.4117004834797
-  },
-  {
-    "date": "2020-08-12T04:05:00",
-    "unit": "mg/dL",
-    "amount": 85.34467074869607
-  },
-  {
-    "date": "2020-08-12T04:10:00",
-    "unit": "mg/dL",
-    "amount": 85.30311598491548
-  },
-  {
-    "date": "2020-08-12T04:15:00",
-    "unit": "mg/dL",
-    "amount": 85.28537663283302
-  },
-  {
-    "date": "2020-08-12T04:20:00",
-    "unit": "mg/dL",
-    "amount": 85.28556920063926
-  },
-  {
-    "date": "2020-08-12T04:25:00",
-    "unit": "mg/dL",
-    "amount": 85.28653120549029
-  },
-  {
-    "date": "2020-08-12T04:30:00",
-    "unit": "mg/dL",
-    "amount": 85.28677865928194
-  },
-  {
-    "date": "2020-08-12T04:35:00",
-    "unit": "mg/dL",
-    "amount": 85.2867985741656
-  }
-]
